### 2020-06-26 #2397 @whedon
[PRE REVIEW]: `swiftsimio`: A python library for reading SWIFT data
**Submitting author:** @JBorrow (<a href="http://orcid.org/0000-0002-1327-1921">Josh Borrow</a>)
**Repository:** <a href="https://github.com/swiftsim/swiftsimio" target ="_blank">https://github.com/swiftsim/swiftsimio</a>
**Version:** v4.0.0
