### 2020-06-26 #2397 @JBorrow
Ah - apologies - I had tried to use the yaml metadata at the top... But must have done something wrong.

The submission is at https://github.com/swiftsim/swiftsimio and docs at swiftsimio.readthedocs.org.
