### 2020-06-26 #2397 @kthyng
Ok, I updated the site above.

@openjournals/dev at this pre-review stage, does this repo name need to be updated in the system so that it is properly used in the review?
