### 2020-06-26 #2397 @arfon
@JBorrow - could you move the contents of this repository (https://github.com/SWIFTSIM/swiftsimio-joss-paper) into https://github.com/swiftsim/swiftsimio. It can be in a separate branch if you like.
