### 2020-06-28 #2397 @arfon
Also, @JBorrow - could you tell me if this submission is associated with a paper in a AAS journal?
