### 2020-06-29 60ad457
Added JOSS paper v1.0
