+++ paper.md
@@ -1,0 +1,178 @@
+---
+title: "`swiftsimio`: A python library for reading SWIFT data"
+tags:
+  - Python
+  - astronomy
+  - cosmology
+  - simulations
+  - i/o
+authors: 
+  - name: Josh Borrow
+    orcid: 0000-0002-1327-1921
+    affiliation: 1
+  - name: Alexei Borrisov
+    orcid: 0000-0001-6247-471X
+    affiliation: 2
+affiliations:
+  - name: Institute for Computational Cosmology, Durham University
+    index: 1
+  - name: School of Physics and Astronomy, University of Edinburgh
+    index: 2
+date: 26 June 2020
+codeRepository: https://github.com/swiftsim/swiftsimio
+license: LGPLv3
+bibliography: bibliography.bib
+---
+
+# Summary
+
+`swiftsimio` is a python package for reading data created by the SWIFT [@SWIFT]
+simulation code. SWIFT is designed to run cosmological hydrodynamics
+simulations that produce petabytes of data, and `swiftsimio` leverages the
+custom metadata that SWIFT produces to allow for chunked loading of the
+particle data and to enable integration with `unyt` [@unyt].
+
+# Background
+
+Cosmological gallaxy formation simulations are used to track simulated galaxies
+across cosmic time in an attempt to better understand the process of galaxy
+formation and evolution. As time has progressed, so has the scale of these
+simulations. The state-of-the-art simulations planned for the next decade,
+using codes like SWIFT, will generate petabytes of data thanks to their use of
+hundreds of billions of particles.  Analysing this data presents a unique
+challenge, with the data reduction performed on either single compute nodes or
+on individual desktop machines. 
+
+In the original EAGLE simulation [@EAGLE], just the co-ordinates of the gas
+particles in a single snapshot used 82 Gb of storage space. State-of-the-art
+simulations are now using at least 10 times as many particles as this, making
+an analysis pipeline that reads the whole array each time expensive in the best
+case and infeasible in the worst. There are two useful properties of this data:
+it is stored in the HDF5 format [@hdf5], allowing for easy slicing, and usually users
+are interested in a very small sub-set of the data, usually less than 1%, at a
+time.
+
+This requirement to load less than 1% of the data at a time is primarily due to
+the huge dynamic range present in these simulations [@dynamicRange]. Although
+the simulation volume may be hundreds of megaparsecs on a side, the objects
+that form under self-gravity are typically less than a few megaparsecs in
+diameter, representing a very small proportion of the total volume. Users are
+usually interested in using the particle data present in a few objects
+(selected using pre-computed data in 'halo catalogues') at any given time.
+
+# Structure of a SWIFT snapshot
+
+At pre-determined times during a SWIFT simulation, a full particle dump is
+performed. This particle dump is stored in the HDF5 format, with arrays
+corresponding to different particle properties. The overall structure of the
+file conforms to the Gadget-2 specification [@Gadget2]. This is to enable
+compatibility with previous analysis pipelines.
+
+The SWIFT snapshot files have main datasets called `PartType{0,1,2,3,4,5}`,
+each with sub-datasets such as `Coordinates`, `Velocities`, etc.  that contain
+the co-ordinates and velocities for particles of that type respectively. When the
+file is opened in `swiftsimio`, this is translated to an object hierarchy, so that
+for a user to access the co-ordinates of the gas particles they would use code
+similar to:
+```python
+from swiftsimio import load
+
+data = load("/path/to/snapshot.hdf5")
+coordinates = data.gas.coordinates
+```
+where here `coordinates` is an `unyt` array containing the co-ordinates of all
+of the gas particles in the file with appropriate units and cosmology information
[…]
