+++ bibliography.bib
@@ -1,0 +1,126 @@
+
+@article{yt,
+	title = {yt: {A} {MULTI}-{CODE} {ANALYSIS} {TOOLKIT} {FOR} {ASTROPHYSICAL} {SIMULATION} {DATA}},
+	volume = {192},
+	issn = {0067-0049, 1538-4365},
+	shorttitle = {yt},
+	url = {https://iopscience.iop.org/article/10.1088/0067-0049/192/1/9},
+	doi = {10.1088/0067-0049/192/1/9},
+	abstract = {The analysis of complex multiphysics astrophysical simulations presents a unique and rapidly growing set of challenges: reproducibility, parallelization, and vast increases in data size and complexity chief among them. In order to meet these challenges, and in order to open up new avenues for collaboration between users of multiple simulation platforms, we present yt (available at http://yt.enzotools.org/) an open source, communitydeveloped astrophysical analysis and visualization toolkit. Analysis and visualization with yt are oriented around physically relevant quantities rather than quantities native to astrophysical simulation codes. While originally designed for handling Enzo’s structure adaptive mesh reﬁnement data, yt has been extended to work with several different simulation methods and simulation codes including Orion, RAMSES, and FLASH. We report on its methods for reading, handling, and visualizing data, including projections, multivariate volume rendering, multidimensional histograms, halo ﬁnding, light cone generation, and topologically connected isocontour identiﬁcation. Furthermore, we discuss the underlying algorithms yt uses for processing and visualizing data, and its mechanisms for parallelization of analysis tasks.},
+	language = {en},
+	number = {1},
+	urldate = {2020-04-17},
+	journal = {The Astrophysical Journal Supplement Series},
+	author = {Turk, Matthew J. and Smith, Britton D. and Oishi, Jeffrey S. and Skory, Stephen and Skillman, Samuel W. and Abel, Tom and Norman, Michael L.},
+	month = jan,
+	year = {2011},
+	pages = {9},
+	file = {Turk et al. - 2011 - yt A MULTI-CODE ANALYSIS TOOLKIT FOR ASTROPHYSICA.pdf:/Users/mphf18/Zotero/storage/NRC9WYWY/Turk et al. - 2011 - yt A MULTI-CODE ANALYSIS TOOLKIT FOR ASTROPHYSICA.pdf:application/pdf}
+}
+
+@article{Gadget2,
+	title = {The cosmological simulation code gadget-2},
+	volume = {364},
+	issn = {0035-8711, 1365-2966},
+	url = {https://academic.oup.com/mnras/article-lookup/doi/10.1111/j.1365-2966.2005.09655.x},
+	doi = {10.1111/j.1365-2966.2005.09655.x},
[…]
