+++ README.md
@@ -1,0 +1,5 @@
+`swiftsimio` JOSS Paper
+=======================
+
+Paper accompanying [swiftsimio](https://github.com/swiftsim/swiftsimio)
+for submission to the Journal of Open Source Software (JOSS).
