### 2020-06-29 #2397 @JBorrow
I have moved the paper inside the main repo into the `joss-paper` branch.
