### 2020-06-29 #2397 @JBorrow
> Also, @JBorrow - could you tell me if this submission is associated with a paper in a AAS journal?

This is not associated with any submission in an AAS journal, no - it is a standalone JOSS submission.
