### 2020-06-29 6aa4cae
Update bibliography to have dates
HDF5 (1997-NNNN to 1997)
Pnbody to (2020) as it does not have an official citation.
