+++ bibliography.bib
@@ -44,7 +44,7 @@
   year = 2013
 }
 
-@misc{pnbody, title={pNbody}, url={http://obswww.unige.ch/~revaz/pNbody/index.html}, journal={pNbody 4.0}, author={Revaz, Yves}}
+@misc{pnbody, title={pNbody}, url={http://obswww.unige.ch/~revaz/pNbody/index.html}, journal={pNbody 4.0}, author={Revaz, Yves}, year={2020}}
 
 @book{h5py,
   title={Python and HDF5: Unlocking Scientific Data},
@@ -121,6 +121,6 @@
 @ONLINE{hdf5,
     author = {{The HDF Group}},
     title = "{Hierarchical Data Format, version 5}",
-    year = {1997-NNNN},
+    year = {1997},
     note = {http://www.hdfgroup.org/HDF5/}
 }
