+++ bibliography.bib
@@ -44,7 +44,19 @@
   year = 2013
 }
 
-@misc{pnbody, title={pNbody}, url={http://obswww.unige.ch/~revaz/pNbody/index.html}, journal={pNbody 4.0}, author={Revaz, Yves}, year={2020}}
+@MISC{pnbody,
+       author = {{Revaz}, Yves},
+        title = "{pNbody: A python parallelized N-body reduction toolbox}",
+     keywords = {Software},
+         year = 2013,
+        month = feb,
+          eid = {ascl:1302.004},
+        pages = {ascl:1302.004},
+archivePrefix = {ascl},
+       eprint = {1302.004},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013ascl.soft02004R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
 @book{h5py,
   title={Python and HDF5: Unlocking Scientific Data},
