### 2020-06-29 #2397 @JBorrow
Thanks, I've updated them. I used the 'official' HDF citation but this had 1997-NNNN in it's date field, I have changed this to just be 1997.

I've changed pnbody to use their ASCL citation.
