### 2020-07-03 #2397 @JBorrow
Great, thanks! As this is generally a post-processing code I am happy to provide some more exciting snapshots for the reviewers to use, should they wish. They can also alternatively use the data downloaded for the unit tests, but this is quite boring (essentially a uniform box).
