### 2020-07-03 #2430 @whedon
[REVIEW]: `swiftsimio`: A python library for reading SWIFT data
**Submitting author:** @JBorrow (<a href="http://orcid.org/0000-0002-1327-1921">Josh Borrow</a>)
**Repository:** <a href="https://github.com/swiftsim/swiftsimio" target ="_blank">https://github.com/swiftsim/swiftsimio</a>
**Version:** v4.0.4
