### 2020-07-08 SWIFTSIM/swiftsimio#64 @dmentipl
[JOSS review] some fixes required in the documentation
JOSS review issue: https://github.com/openjournals/joss-reviews/issues/2430

# Summary

First of all, swiftsimio is a really well written Python package, filling a need to analyse and visualise large SWIFT datasets. The particle masking feature is particularly cool. It has excellent internal documentation via docstrings. I like that you use black for standardised formatting.

There are a few relatively minor issues that need addressing detailed below.

# Code

One minor thing is that there is a slightly confusing message: "You do not have numba installed. Please consider installing if you are going to be doing visualisation or indexing large arrays (pip install numba)" when importing. However, I do have numba (v0.48.0) installed. I think `numba.core`, which is the missing import that generates this message, was added in v0.49.1.

I have a small suggestion that it would be useful to add an `environment.yml` file to set up a conda environment for users or for development. E.g. I used this while testing the documentation code examples:

```yaml
name: swiftsimio

channels:
  - conda-forge

dependencies:
  - python>=3.6
  - black
  - h5py
  - ipython
  - matplotlib
  - numba>=0.50
  - numpy
  - pip
  - pykdtree
  - pytest
  - scipy
  - tqdm
  - unyt
  - pip:
    - h5glance
    - py-sphviewer
    - swiftsimio
    - velociraptor
    - wily
```

# Manuscript

The only comments I have about the manuscript are:

- there is a typo in the "Background" section in paragraph 1: "gallaxy";
- a small style-point: I think Python should be capitalised.

# Documentation

There are several (mostly minor) issues with the documentation on readthedocs. (I was working off the stable version v4.0.0; commit hash `16e79ce8`, using the provided dataset from [http://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/cosmological_volume.hdf5](http://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/cosmological_volume.hdf5).)

There are two changes that I suggest are important additions to the documentation.

1. I would like to see the output figures generated by the code examples in the "Visualisation" sections to check if the code is working when I run it.
2. It would be good to include a link to an example "catalogue file" required in the code examples in the "Visualisation/Projection" page.

In what follows I've detailed issues I found while running the example code in the documentation.

## Loading Data

### Using metadata

- Where it states "This will output [4.38929954e+26 4.38929954e+26 4.38929954e+26] cm" instead I get: `[142.24751067 142.24751067 142.24751067] Mpc`.
- There is a typo in "(a :mod:datetime object describing when the snapshot was written to disk)". I think you just need to replace ":mod:" with ":py:mod:".

### Reading particle data

- Where it states "will output 3.08567758e+24*cm." instead I get: `Mpc`.
- Where it states "which will output '\\frac{M_\\odot}{\\rm{Mpc}^{3}}'." instead I get: `2.0 \times 10^{30}\ \frac{\rm{kg}}{\rm{Mpc}^{3}}`.

### Named columns

- I think this is not formatted as intended. There seems to be a syntax error. I couldn't really run the code.

### Non-unyt properties

- Where it states "which will output Co-moving mass densities of the particles." instead I get: `AttributeError: 'cosmo_array' object has no attribute 'description'`.

## Masking

### Spatial-only masking

- In the example, I think `data = load(filename, mask=mask)` should be `sw.data = load(filename, mask=mask)`.

### Full mask

- In the example on the line containing `mask.constrain_mask("gas", "density", 0.4 * density_units, 0.8 * density_units)` I get `AttributeError: 'SWIFTUnits' object has no attribute 'gas'`.

### Writing a subset of snapshot

- In the example on the line containing `sw.subset_writer.write_subset("test_subset.hdf5", mask)` I get `AttributeError: module 'swiftsimio' has no attribute 'subset_writer'`.

## Visualisation — Projection

It would be good to include the images generated by the examples, preferably using the available dataset.

### Rotation

- In the example, a catalogue file is required. Can this be provided?

### Lower-level API

- It would be nice if the example was executable.

## Visualisation — Slices

### Example

- In the second example, should "project_gas" should be replaced by "slice_gas"?

## Visualisation — Volume Rendering

### Example

- In the first code block, should "slice_gas" be replaced by "render_gas"?
- It seems like there may be a memory leak, or at least my laptop doesn't have enough memory to run the second code block with the dataset provided. The code crashes IPython.

## Command-line Utilities

- Running `swiftsnap` on the provided dataset gives me: `AttributeError: 'str' object has no attribute 'decode'`
