### 2020-07-08 SWIFTSIM/swiftsimio#64 @JBorrow
Thank you for such a detailed review. I will respond to your comments with version 4.0.1. Apologies that the sample data was so out of date - we never really use that so it just got forgotten about it.
