### 2020-07-08 31e192f
Added backwards compatibility with older versions of numba
