+++ swiftsimio/accelerated.py
@@ -11,21 +11,25 @@ from typing import Tuple
 try:
     from numba import jit, prange
     from numba.core.config import NUMBA_NUM_THREADS as NUM_THREADS
-except ImportError:
-    print(
-        "You do not have numba installed. Please consider installing "
-        "if you are going to be doing visualisation or indexing large arrays "
-        "(pip install numba)"
-    )
+except (ImportError, ModuleNotFoundError):
+    try:
+        from numba import jit, prange
+        from numba.config import NUMBA_NUM_THREADS as NUM_THREADS
+    except (ImportError, ModuleNotFoundError):
+        print(
+            "You do not have numba installed. Please consider installing "
+            "if you are going to be doing visualisation or indexing large arrays "
+            "(pip install numba)"
+        )
 
-    def jit(*args, **kwargs):
-        def x(func):
-            return func
+        def jit(*args, **kwargs):
+            def x(func):
+                return func
 
-        return x
+            return x
 
-    prange = range
-    NUM_THREADS = 1
+        prange = range
+        NUM_THREADS = 1
 
 
 @jit(nopython=True)
