### 2020-07-08 78983ea
Reading particle metadata changes
