+++ swiftsimio/masks.py
@@ -197,26 +197,32 @@ class SWIFTMask(object):
 
         current_mask = getattr(self, ptype)
 
-        handle = {v: k for k, v in getattr(metadata.particle_fields, ptype).items()}[
-            quantity
-        ]
-        unit = getattr(self.units, ptype)[quantity]
-        # We use the type and not the number because it is far easier for users to understand.
-        particle_number = {
-            v: k for k, v in metadata.particle_types.particle_name_underscores.items()
-        }[ptype]
+        particle_metadata = getattr(self.metadata, f"{ptype}_properties")
+        unit_dict = {
+            k: v
+            for k, v in zip(
+                particle_metadata.field_names, particle_metadata.field_units
+            )
+        }
+
+        unit = unit_dict[quantity]
+
+        handle_dict = {
+            k: v
+            for k, v in zip(
+                particle_metadata.field_names, particle_metadata.field_paths
+            )
+        }
+
+        handle = handle_dict[quantity]
+
         # Load in the relevant data.
 
         with h5py.File(self.metadata.filename, "r") as file:
             # Surprisingly this is faster than just using the boolean
             # indexing because h5py has slow indexing routines.
             data = unyt.unyt_array(
-                np.take(
-                    file[f"PartType{particle_number}/{handle}"],
-                    np.where(current_mask)[0],
-                    axis=0,
-                ),
-                units=unit,
+                np.take(file[handle], np.where(current_mask)[0], axis=0), units=unit
             )
 
         new_mask = np.logical_and.reduce([data > lower, data <= upper])
@@ -432,7 +438,8 @@ class SWIFTMask(object):
             current_offsets = {}
             if not hasattr(self, "cell_mask"):
                 raise RuntimeError(
-                    "Subset writing requires specifying a cell mask. Please use constrain_spatial with a suitable restrict array to generate one."
+                    "Subset writing requires specifying a cell mask. Please use "
+                    "constrain_spatial with a suitable restrict array to generate one."
                 )
             for part_type, counts in self.counts.items():
                 masked_counts[part_type] = counts * self.cell_mask
