### 2020-07-08 a6b6118
Update documentation with minor fixes
