+++ docs/source/loading_data/index.rst
@@ -19,7 +19,14 @@ To get started, first locate any SWIFT data that you wish to analyse. If you
 don't have any handy, you can always download our test cosmological volume
 at:
 
-``http://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/cosmological_volume.hdf5``
+``http://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/cosmological_volume_example.hdf5``
+
+with associated halo catalogue at
+
+``http://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/cosmological_volume_example.properties``
+
+which is needed should you wish to use the ``velociraptor`` integration library in the
+visualisation examples.
 
 To create your first instance of :obj:`swiftsimio.reader.SWIFTDataset`, you can
 use the helper function :mod:`swiftsimio.load` as follows:
