### 2020-07-08 a290665
Added minor paper changes
