### 2020-07-08 e81cf3c
Added example output images
