### 2020-07-08 ba66fe0
Fixed a huge number of previously un-noticed documentation warnings
