### 2020-07-08 SWIFTSIM/swiftsimio#64 @JBorrow
I believe that #65 should have addressed all of these comments.

Going through one-by-one:

1. I have enabled the code to work with slightly older versions of `numba`
2. Thanks for the `conda` file, we don't use it but I have added this to the repo so that users of that package management software can develop on `swiftsimio` easier.
3. Manuscript changes have been performed in the `joss-paper` branch.
4. I have updated the example data to be able to work with all of the examples; apologies again that this was so out of date.
5. I have added a velociraptor catalogue to the loading data page for use with the snapshot.
6. I have fixed this issue (we used to use cgs for base units everywhere; that's now fixed).
7. I have fixed the reference to `datetime`
8. Same as above; fixed.
9. Same as above; fixed.
10. Named columns should work with the new example dataset; this older data was not compatible with the new named columns scheme.
11. This was changed when we convinced the `unyt` folks to have `name` attribute in their array - we changed from `description` to `name`. This is updated within the documentation now.
12. This typo has been corrected.
13. Thanks - you discovered a bug! We have fixed this now.
14. We were missing an import in `__init__.py`; this should now be fixed.
15. I have added example images in some cases within the visualisation, and provided the catalogue file for download.
16. I can't think of a nice way of making that example executable without having a random field of particles visualised... For now, I would like to keep it as it is.
17. Yes, `project_gas` has been changed to `slice_gas`
18. Yes, I have made this change.
19. Unfortunately no memory leak, but this is very memory-intensive. I have changed the example to use a 256^3 grid instead of 1024^3 (which is 8 Gb and for the parallel version of this we create N_threads copies, oops!)
20. The CLI `swiftsnap` has been fixed and I've added some more safety checks to stop this happening with other quantities.

I would like to thank you again for your kind words and your fast and detailed review!
