### 2020-07-09 SWIFTSIM/swiftsimio#64 @dmentipl
JOSS review issue: https://github.com/openjournals/joss-reviews/issues/2430

> I would like to thank you again for your kind words and your fast and detailed review!

No worries. Happy to help!

Thanks for the changes you've made. That's covered most of the issues.

Unfortunately I can't load the new dataset (at [http://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/cosmo_volume_example.hdf5](http://virgodb.cosma.dur.ac.uk/swift-webstorage/IOExamples/cosmo_volume_example.hdf5)). Running off the swiftsimio master branch, I get the following traceback:

```python
~/conda/envs/swiftsimio/lib/python3.8/site-packages/swiftsimio/reader.py in postprocess_header(self)
    269             try:
--> 270                 self.snapshot_date = datetime.strptime(
    271                     self.header["Snapshot date"].decode("utf-8"), "%H:%M:%S %Y-%m-%d %Z"

~/conda/envs/swiftsimio/lib/python3.8/_strptime.py in _strptime_datetime(cls, data_string, format)
    567     format string."""
--> 568     tt, fraction, gmtoff_fraction = _strptime(data_string, format)
    569     tzname, gmtoff = tt[-2:]

~/conda/envs/swiftsimio/lib/python3.8/_strptime.py in _strptime(data_string, format)
    348     if not found:
--> 349         raise ValueError("time data %r does not match format %r" %
    350                          (data_string, format))

ValueError: time data '13:59:18 2020-07-08 BST' does not match format '%H:%M:%S %Y-%m-%d %Z'

<ipython-input-2-9e3866592fc4> in <module>
----> 1 data = sw.load('cosmo_volume_example.hdf5')

~/conda/envs/swiftsimio/lib/python3.8/site-packages/swiftsimio/__init__.py in load(filename, mask)
     90     """
     91
---> 92     return SWIFTDataset(filename, mask=mask)
     93
     94

~/conda/envs/swiftsimio/lib/python3.8/site-packages/swiftsimio/reader.py in __init__(self, filename, mask)
   1214
   1215         self.get_units()
-> 1216         self.get_metadata()
   1217         self.create_particle_datasets()
   1218

~/conda/envs/swiftsimio/lib/python3.8/site-packages/swiftsimio/reader.py in get_metadata(self)
   1250         """
   1251
-> 1252         self.metadata = SWIFTMetadata(self.filename, self.units)
   1253
   1254         return

~/conda/envs/swiftsimio/lib/python3.8/site-packages/swiftsimio/reader.py in __init__(self, filename, units)
    136         self.get_named_column_metadata()
    137
--> 138         self.postprocess_header()
    139
    140         self.load_particle_types()

~/conda/envs/swiftsimio/lib/python3.8/site-packages/swiftsimio/reader.py in postprocess_header(self)
    276                 # a British (en_GB) machine, and then tried to read on a Dutch
    277                 # machine (nl_NL), this would _not_ work because %c is different.
--> 278                 self.snapshot_date = datetime.strptime(
    279                     self.header["Snapshot date"].decode("utf-8"), "%c\n"
    280                 )

~/conda/envs/swiftsimio/lib/python3.8/_strptime.py in _strptime_datetime(cls, data_string, format)
    566     """Return a class cls instance based on the input string and the
    567     format string."""
--> 568     tt, fraction, gmtoff_fraction = _strptime(data_string, format)
    569     tzname, gmtoff = tt[-2:]
    570     args = tt[:6] + (fraction,)

~/conda/envs/swiftsimio/lib/python3.8/_strptime.py in _strptime(data_string, format)
    347     found = format_regex.match(data_string)
    348     if not found:
--> 349         raise ValueError("time data %r does not match format %r" %
    350                          (data_string, format))
    351     if len(data_string) != found.end():

ValueError: time data '13:59:18 2020-07-08 BST' does not match format '%c\n'
```

A couple of other comments about the docs:

- I suggest replacing "cosmological_volume.hdf5" and "my_snapshot_0000.hdf5" everywhere with "cosmo_volume_example.hdf5" now that you've updated the example dataset. (If that is what is intended.)
- I'm still a bit confused about the "Named columns" section of the "Loading Data" page. The code box does not contain valid Python. Maybe I'm misinterpreting it.
