### 2020-07-09 824725c
Fixed incorrectly-nested try/except statements.
