### 2020-07-09 eeae861
Change name of snapshot in examples
