+++ swiftsimio/__version__.py
@@ -1,1 +1,1 @@
-__version__ = "4.0.1"
+__version__ = "4.0.1hotfix"
