### 2020-07-09 SWIFTSIM/swiftsimio#64 @JBorrow
Thanks, I've hopefully fixed that in version 4.0.2. Sorry about that - we don't have any machines that live in different timezones (with different localisations too) that we can test on easily. I should look into how to do this with travis.

I have replaced the names of all of the files at your request.

I have also changed the Named Columns section of the documentation to be clearer: https://swiftsimio.readthedocs.io/en/latest/loading_data/index.html#named-columns
