### 2020-07-10 SWIFTSIM/swiftsimio#64 @dmentipl
No worries.

Thanks for making those changes. It all looks good.
