+++ swiftsimio/initial_conditions/__init__.py
@@ -1,0 +1,3 @@
+"""
+Initial conditions generation.
+"""
