+++ swiftsimio/initial_conditions/generate_particles.py
@@ -1,0 +1,5 @@
+"""
+Particle generation code.
+
+TBD
+"""
