+++ swiftsimio/initial_conditions/generate_particles.py
@@ -3,3 +3,702 @@ Particle generation code.
 
 TBD
 """
+
+#--------------------------------------------
+# IC generation related stuff
+#--------------------------------------------
+
+
+import numpy as np
+from matplotlib import pyplot as plt
+from mpl_toolkits.mplot3d import Axes3D
+from scipy import stats
+
+from .IC_grid import compute_smoothing_lengths, build_grid
+from .IC_kernel import get_kernel_data
+
+
+
+IC_ITER_MAX = 2000                  # max numbers of iterations for generating IC conditions
+IC_CONVERGENCE_THRESHOLD = 1e-3     # if enough particles are displaced by distance below threshold * mean interparticle distance, stop iterating.
+IC_TOLERANCE_PART = 1e-2            # tolerance for not converged particle fraction: this fraction of particles can be displaced with distances > threshold
+IC_DISPLACEMENT_THRESHOLD = 1e-2    # Don't even thing about stopping until every particle is displaced by distance < threshold * mean interparticle distance.
+                                    # Only after every particle has moved with distance smaller than this threshold, the IC_CONVERGENCE_THRESHOLD is
+                                    # taken into account.
+IC_REDISTRIBUTE_AT_ITERATION = 10   # redistribute a handful of particles every IC_REDISTRIBUTE_AT_ITERATION iteration
+IC_DELTA_REDUCTION_FACTOR = 0.97    # reduce normalization constant for particle motion by this factor after every iteration
+IC_DELTA_INIT = 0.1                 # initial normalization constant for particle motion in units of mean interparticle distance
+IC_DELTA_MIN = 1e-4                 # minimal normalization constant for particle motion in units of mean interparticle distance
+IC_REDISTRIBUTE_FRACTION = 0.01     # fraction of particles to redistribute when doing so
+IC_NO_REDISTRIBUTION_AFTER = 200    # don't redistribute particles after this iteration
+IC_PLOT_AT_REDISTRIBUTION = True    # create and store a plot of the current situation before redistributing?
+
+
+np.random.seed(666)
+
+
+
+def IC_generation_set_params(
+            iter_max                    = IC_ITER_MAX,
+            convergence_threshold       = IC_CONVERGENCE_THRESHOLD,
+            tolerance_part              = IC_TOLERANCE_PART,
+            displacement_threshold      = IC_DISPLACEMENT_THRESHOLD,
+            redistribute_at_iteration   = IC_REDISTRIBUTE_AT_ITERATION,
+            delta_init                  = IC_DELTA_INIT,
+            delta_reduction_factor      = IC_DELTA_REDUCTION_FACTOR,
+            delta_min                   = IC_DELTA_MIN,
+            redistribute_fraction       = IC_REDISTRIBUTE_FRACTION,
+            no_redistribution_after     = IC_NO_REDISTRIBUTION_AFTER,
+            plot_at_redistribution      = IC_PLOT_AT_REDISTRIBUTION
+        ):
+    """
+    Change the global IC generation iteration parameters.
+
+    parameters:
+
+        <int>   iter_max                    max numbers of iterations for generating IC conditions
+        <float> convergence_threshold       if enough particles are displaced by distance below 
+                                                threshold * mean interparticle distance, stop iterating.
+        <float> tolerance_part              tolerance for not converged particle fraction: 
+                                                this fraction of particles can be displaced with 
+                                                distances > threshold
[…]
