+++ swiftsimio/initial_conditions/__init__.py
@@ -1,3 +1,5 @@
 """
 Initial conditions generation.
 """
+
+from .generate_particles import *
