+++ swiftsimio/initial_conditions/IC_grid.py
@@ -1,0 +1,484 @@
+#!/usr/bin/env python3
+
+
+#----------------------------------------
+# Grid related functions and classes
+#----------------------------------------
+
+
+import numpy as np
+import copy
+
+
+from particle_hydro_kernel import get_kernel_data
+
+min_neighbour_fact_for_cells = 2 # require minimum of min_neighbour_fact_for_cells * nngb particles in cells for h iteration
+ITER_MAX = 100 # max number of iterations to do
+EPSILON = 1e-3 # convergence tolerance
+
+
+
+
+
+
+def compute_smoothing_lengths(x, m, eta, kernel='cubic spline', ndim=2, periodic=True, ncells = None):
+    """
+    Compute the smoothing length and density of particles with given positions x and masses m
+        x:      particle coordinates
+        m:      particle masses
+        eta:    "resolution", that defines number of neighbours
+        kernel: which kernel to use
+        ndim:   number of dimensions
+        periodic: whether the gig is periodic
+        ncells: a first guess for how many cells in every dimension you want
+
+    returns:
+        h:      1D numpy array containing particle smoothing lengths
+        rho:    1D numpy array containing particle densities computed with computed smoothing lengths
+        grid:   list of cells constituting the grid
+        ncells: number of cells in each dimension
+        neighbours: list of arrays containing neighbour particle indices for each particle
+    """
+
+    grid, ncells = build_grid(x, m, eta, kernel=kernel, ndim=ndim, periodic=periodic, verbose=False, ncells=ncells)
+
+    nparttot = m.shape[0]
+
+    h = np.zeros(nparttot, dtype=np.float)
+    rho = np.zeros(nparttot, dtype=np.float)
+    neighbours = [[] for i in range(nparttot)]
+
+    kernel_func, kernel_derivative, kernel_gamma = get_kernel_data(kernel, ndim)
+
+    if ndim == 1:
+        nngb = 2 * eta * kernel_gamma
+    elif ndim == 2:
+        nngb = np.pi * (eta * kernel_gamma)**2
+
+    HMAX = 0.5
+    HMIN = nparttot**(-1./ndim) * 2 # take at least 2 interparticle distances
+
+    # get a rounded up integer to use in arrays
+    nngb = int(nngb + 0.5)
+    
+
+    for c in grid:
+        nbrs = c.get_neighbours(ncells, ndim, periodic=periodic)
+        # get all particles in this neighbourhood
+        allparts = []
+        for n in nbrs:
+            allparts += grid[n].parts
+
+        # get all neighbour particle positions and masses 
+        xn = x[allparts]
+        mn = m[allparts]
+
+        # now loop over every particle in this cell
+        for p in c.parts:
+            if ndim == 1:
+                r = xn - x[p]
+                if (periodic):
+                    r[r>0.5] -= 1
+                    r[r<-0.5] += 1
+                r = np.abs(r)
+            elif ndim == 2:
+                dx = xn[:,0] - x[p,0]
+                if (periodic):
+                    dx[dx>0.5] -= 1
+                    dx[dx<-0.5] += 1
+                dy = xn[:,1] - x[p,1]
+                if (periodic):
+                    dy[dy>0.5] -= 1
+                    dy[dy<-0.5] += 1
+                r = np.sqrt(dx**2 + dy**2)
+    
+            sortind = np.argsort(r)
+            r = r[sortind]
+            msort = mn[sortind]
+            nbrsort = np.array(allparts)[sortind]
+
+            # prepare iterations
+            niter = 0
+            Hi = r[nngb] # initial guess
+            Hiprev = 0. # store previous one
+
+            # start iterating
+            while True:
+                niter += 1
+
+                i = 0
+                while r[i] <= Hi: # find where to stop
+                    i += 1
+                    if i == len(r): 
[…]

[…]
