### 2020-07-23 fbdfd6b
moved globals to a dict
