### 2020-07-23 292b2fe
added 3D kernels
