+++ swiftsimio/initial_conditions/IC_kernel.py
@@ -9,9 +9,16 @@ import numpy as np
 
 
 
-def W_cubic_spline(r, H, ndim = 2):
+def W_cubic_spline(r, H):
     """
     Cubic spline kernel.
+    Parameters:
+        r:  numpy array of distances between particles.
+        H:  compact suppor radius of kernel. Scalar or numpy array of same shape as r.
+
+    returns:
+        W:  numpy array of evaluated kernels with same shape as r. It's however still
+            un-normalized and needs to be divided by H^ndim to get the proper values.
     """
     if type(r) != np.ndarray:
         r = np.array(r)
@@ -19,66 +26,59 @@ def W_cubic_spline(r, H, ndim = 2):
     q = r / H
     W = np.zeros(q.shape)
 
-    W[q<=1] += (1. - q[q<=1])**3
-    W[q<=0.5] -= 4*(0.5 - q[q<0.5])**3
     #  if q <= 1.:
     #      W += (1. - q)**3
     #  if q <= 0.5:
     #      W -= 4*(0.5 - q)**3
-
-    if ndim == 1:
-        W /= H
-        W *= kernel_norm_1D['cubic spline']
-    elif ndim == 2:
-        W /= H**2
-        W *= kernel_norm_2D['cubic spline']
+    W[q<=1] += (1. - q[q<=1])**3
+    W[q<=0.5] -= 4*(0.5 - q[q<0.5])**3
 
     return W
 
 
 
 
-def dWdr_cubic_spline(r, H, ndim = 2):
+def dWdr_cubic_spline(r, H):
     """
-    Cubic spline kernel derivative
-    """
+    Cubic spline kernel derivative.
+    Parameters:
+        r:  numpy array of distances between particles.
+        H:  compact suppor radius of kernel. Scalar or numpy array of same shape as r.
 
-    if type(r) != np.ndarray:
-        r = np.array(r)
+    returns:
+        dWdr:  numpy array of evaluated kernel derivatives with same shape as r. It's
+                however un-normalized and needs to be divided by H^(ndim+1) to get the proper values.
+    """
 
     q = r / H
     dW = np.zeros(q.shape)
 
-    dW[q<=1] -= 3 * (1. - q[q<=1])**2
-    dW[q<=0.5] += 12 * (0.5 - q[q<=0.5])**2
     #  if q <= 1.:
     #      W -= 3 * (1. - q)**2
     #  if q <= 0.5:
     #      W += 12*(0.5 - q)**2
-
-
-    if ndim == 1:
-        dW /= H**2
-        dW *= kernel_norm_1D['cubic spline']
-    elif ndim == 2:
-        dW /= H**3
-        dW *= kernel_norm_2D['cubic spline']
+    dW[q<=1] -= 3 * (1. - q[q<=1])**2
+    dW[q<=0.5] += 12 * (0.5 - q[q<=0.5])**2
 
     return dW
 
 
+# dictionary "pointing" to the correct functions to call
+kernel_funcs = {}
+kernel_funcs['cubic spline'] = W_cubic_spline
+
+kernel_derivatives = {}
+kernel_derivatives['cubic spline'] = dWdr_cubic_spline
+
+
+
+# Constants are from Dehnen & Aly 2012
 
 kernel_gamma_1D = {}
 kernel_gamma_1D['cubic spline'] = 1.732051
 
 kernel_norm_1D = {}
 kernel_norm_1D['cubic spline'] = 2.666667
-
-kernel_func_1D = {}
-kernel_func_1D['cubic spline'] = W_cubic_spline
-
-kernel_derivative_1D = {}
-kernel_derivative_1D['cubic spline'] = dWdr_cubic_spline
 
[…]
