### 2020-07-23 769d14c
added quick kernel test
