### 2020-07-24 a900d86
x is now 3D array always. Differentiating between icSimParams and icRunParams dicts.
