+++ swiftsimio/initial_conditions/generate_particles.py
@@ -16,13 +16,37 @@ from scipy import stats
 from .IC_grid import compute_smoothing_lengths, build_grid
 from .IC_kernel import get_kernel_data
 
+from swiftsimio.optional_packages import KDTree, TREE_AVAILABLE
 
 
 np.random.seed(666)
 
 
+def IC_set_IC_params(
+            boxsize     = [1, 1], 
+            periodic    = True,
+            nx          = 100,
+            ndim        = 2
+        ):
+    """
+    Change the initial conditions parameters.
 
-def IC_generation_set_params(
+    TODO: docs
+
+    returns:
+        dict with these parameters.
+    """
+
+    icSimParams = {}
+    icSimParams['boxsize'] = boxsize
+    icSimParams['periodic'] = periodic
+    icSimParams['nx'] = nx
+    icSimParams['ndim'] = ndim
+
+    return icSimParams
+
+
+def IC_set_run_params(
             iter_max                    = 2000,
             convergence_threshold       = 1e-3,
             tolerance_part              = 1e-2,
@@ -33,7 +57,9 @@ def IC_generation_set_params(
             redistribute_frequency      = 20,
             redistribute_fraction       = 0.01,
             no_redistribution_after     = 200,
-            plot_at_redistribution      = True
+            plot_at_redistribution      = True, 
+            kernel                      = 'cubic spline',
+            eta                         = 1.2348
         ):
     """
     Change the global IC generation iteration parameters.
@@ -66,45 +92,67 @@ def IC_generation_set_params(
         dictionnary containing all this data.
     """
 
-    icparams = {}
+    icRunParams = {}
 
-    icparams['ITER_MAX'] = iter_max
-    icparams['CONVERGENCE_THRESHOLD'] = convergence_threshold
-    icparams['TOLERANCE_PART'] = tolerance_part
-    icparams['DISPLACEMENT_THRESHOLD'] = displacement_threshold
-    icparams['DELTA_INIT'] = delta_init
-    icparams['DELTA_REDUCTION_FACTOR'] = delta_reduction_factor
-    icparams['DELTA_MIN'] = delta_min
-    icparams['REDISTRIBUTE_FREQUENCY'] = redistribute_frequency
-    icparams['REDISTRIBUTE_FRACTION'] = redistribute_fraction
-    icparams['NO_REDISTRIBUTION_AFTER'] = no_redistribution_after
-    icparams['PLOT_AT_REDISTRIBUTION'] = plot_at_redistribution
+    icRunParams['ITER_MAX'] = iter_max
+    icRunParams['CONVERGENCE_THRESHOLD'] = convergence_threshold
+    icRunParams['TOLERANCE_PART'] = tolerance_part
+    icRunParams['DISPLACEMENT_THRESHOLD'] = displacement_threshold
+    icRunParams['DELTA_INIT'] = delta_init
+    icRunParams['DELTA_REDUCTION_FACTOR'] = delta_reduction_factor
+    icRunParams['DELTA_MIN'] = delta_min
+    icRunParams['REDISTRIBUTE_FREQUENCY'] = redistribute_frequency
+    icRunParams['REDISTRIBUTE_FRACTION'] = redistribute_fraction
+    icRunParams['NO_REDISTRIBUTION_AFTER'] = no_redistribution_after
+    icRunParams['PLOT_AT_REDISTRIBUTION'] = plot_at_redistribution
+    icRunParams['KERNEL'] = kernel
+    icRunParams['ETA'] = eta
 
-    return icparams
+    return icRunParams
 
 
 
-def IC_uniform_coordinates(nx, ndim = 2, periodic = True):
+def IC_uniform_coordinates(nx, ndim = 2, boxsize = [1, 1]):
     """
[…]
