### 2020-07-24 3f1c01b
using scipy cKDTree now for smoothing lengths and neighbour searches.
