+++ swiftsimio/initial_conditions/generate_particles.py
@@ -13,7 +13,6 @@ from matplotlib import pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 from scipy import stats
 
-from .IC_grid import compute_smoothing_lengths, build_grid
 from .IC_kernel import get_kernel_data
 
 from swiftsimio.optional_packages import KDTree, TREE_AVAILABLE
@@ -326,8 +325,6 @@ def generate_IC_for_given_density(rho_anal, icSimParams=IC_set_IC_params(), icRu
 
 
 
-    delta_r = np.zeros(x.shape, dtype=np.float)
-
     kernel_func, kernel_derivative, kernel_gamma = get_kernel_data(icRunParams['KERNEL'], ndim)
 
     # expected number of neighbours
@@ -338,22 +335,32 @@ def generate_IC_for_given_density(rho_anal, icSimParams=IC_set_IC_params(), icRu
     elif ndim == 3:
         Nngb = 4 / 3 * np.pi * (kernel_gamma * icRunParams['ETA'])**3
 
+    # round it up for cKDTree
+    Nngb_int = int(Nngb + 0.5)
+
     MID = 1./npart**(1./ndim) # mean interparticle distance
     delta_r_norm = icRunParams['DELTA_INIT'] * MID
     delta_r_min = icRunParams['DELTA_MIN'] * MID
 
+    # empty arrays
+    delta_r = np.zeros(x.shape, dtype=np.float)
+    rho = np.zeros(npart, dtype=np.float)
+    h = np.zeros(npart, dtype=np.float)
 
     if icRunParams['PLOT_AT_REDISTRIBUTION']:
         # drop a first plot
-        h, rho, _, ncells_proper, _ = compute_smoothing_lengths(x, m, icRunParams['ETA'], 
-                    ndim=ndim, periodic=periodic, ncells=None)
+        tree = KDTree(x[:,:ndim], boxsize=boxsize)
+        for p in range(npart):
+            dist, neighs = tree.query(x[p, :ndim], k=Nngb_int)
+            h[p] = dist[-1] / kernel_gamma
+            W = kernel_func(dist, dist[-1])
+            rho[p] = (W * m[neighs]).sum()
         IC_plot_current_situation(True, 0, x, rho, rho_anal, ndim=ndim)
 
 
     # start iteration loop
     iteration = 0
-    ncells = None
-    ncells_proper = None
+
     while True:
 
         iteration += 1
@@ -361,13 +368,23 @@ def generate_IC_for_given_density(rho_anal, icSimParams=IC_set_IC_params(), icRu
         # get analytical density at particle positions
         rhoA = rho_anal(x)
 
+        # reset displacements
+        delta_r.fill(0.)
+
         # re-distribute particles?
         if iteration % icRunParams['REDISTRIBUTE_FREQUENCY'] == 0:
 
             # do we need to compute current densities and h's?
             if icRunParams['PLOT_AT_REDISTRIBUTION'] or icRunParams['NO_REDISTRIBUTION_AFTER'] >= iteration:
-                h, rho, _, ncells_proper, _ = compute_smoothing_lengths(x, m, icRunParams['ETA'], 
-                            kernel=icRunParams['KERNEL'], ndim=ndim, periodic=periodic, ncells=ncells_proper)
+
+                # TODO: move this inside redistribute_particles?
+                # first build new tree
+                tree = KDTree(x[:,:ndim], boxsize=boxsize)
+                for p in range(npart):
+                    dist, neighs = tree.query(x[p, :ndim], k=Nngb_int)
+                    h[p] = dist[-1] / kernel_gamma
+                    W = kernel_func(dist, dist[-1])
[…]
