### 2020-07-25 2049912
naive coordinate generation now works with unyt
