+++ swiftsimio/initial_conditions/generate_particles.py
@@ -8,7 +8,7 @@ Generate particle initial conditions that follow some density function.
 
 
 import numpy as np
-import unyt
+from unyt import unyt_array
 from math import erf
 from matplotlib import pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
@@ -23,10 +23,11 @@ np.random.seed(666)
 
 
 def IC_set_IC_params(
-            boxsize     = unyt.unyt_array([1., 1., 1.], "cm"),
+            boxsize     = unyt_array([1., 1., 1.], "cm"),
             periodic    = True,
             nx          = 100,
-            ndim        = 2
+            ndim        = 2,
+            coordinate_units = None
         ):
     """
     Change the initial conditions parameters.
@@ -42,6 +43,10 @@ def IC_set_IC_params(
     icSimParams['periodic'] = periodic
     icSimParams['nx'] = nx
     icSimParams['ndim'] = ndim
+    if coordinate_units is None:
+        icSimParams['units'] = boxsize.units
+    else:
+        isCimParams['units'] = coordinate_units
 
     return icSimParams
 
@@ -112,7 +117,7 @@ def IC_set_run_params(
 
 
 
-def IC_uniform_coordinates(nx, ndim = 2, boxsize = [1, 1]):
+def IC_uniform_coordinates(icSimParams=IC_set_IC_params()):
     """
     Get the coordinates for a uniform particle distribution.
     nx:         number of particles in each dimension
@@ -122,30 +127,33 @@ def IC_uniform_coordinates(nx, ndim = 2, boxsize = [1, 1]):
         x: np.array((nx**ndim, 3), dtype=float) of coordinates
     """
 
+    nx = icSimParams['nx']
+    boxsize = icSimParams['boxsize']
+    ndim = icSimParams['ndim']
+
     npart = nx**ndim
-    x = np.zeros((npart, 3), dtype=np.float)
+    x = unyt_array(np.zeros((npart, 3), dtype=np.float), boxsize.units)
 
-    # TODO: fix boxlen here
-
-    dxhalf = 0.5/nx
-    dyhalf = 0.5/nx
-    dzhalf = 0.5/nx
+    dxhalf = 0.5 * boxsize[0]/nx
+    dyhalf = 0.5 * boxsize[1]/nx
+    dzhalf = 0.5 * boxsize[2]/nx
 
     if ndim == 1:
-        x[:,0] = np.linspace(dxhalf, 1-dxhalf, nx)
+        x[:,0] = np.linspace(dxhalf, boxsize[0]-dxhalf, nx)
 
     elif ndim == 2:
-        xcoords = np.linspace(dxhalf, 1-dxhalf, nx)
-        ycoords = np.linspace(dyhalf, 1-dyhalf, nx)
+        xcoords = np.linspace(dxhalf, boxsize[0]-dxhalf, nx)
+        ycoords = np.linspace(dyhalf, boxsize[1]-dyhalf, nx)
         for i in range(nx):
             start = i*nx
             stop = (i+1)*nx
             x[start:stop, 0] = xcoords
             x[start:stop, 1] = ycoords[i]
+
     elif ndim == 3:
-        xcoords = np.linspace(dxhalf, 1-dxhalf, nx)
-        ycoords = np.linspace(dyhalf, 1-dyhalf, nx)
-        zcoords = np.linspace(dzhalf, 1-dzhalf, nx)
+        xcoords = np.linspace(dxhalf, boxsize[0]-dxhalf, nx)
+        ycoords = np.linspace(dyhalf, boxsize[1]-dyhalf, nx)
+        zcoords = np.linspace(dzhalf, boxsize[2]-dzhalf, nx)
[…]
