### 2020-07-26 1b572c1
bugfixes and cleanups
