+++ swiftsimio/initial_conditions/generate_particles.py
@@ -27,7 +27,8 @@ def IC_set_IC_params(
             periodic    = True,
             nx          = 100,
             ndim        = 2,
-            coordinate_units = None
+            unit_l      = None,
+            unit_m      = 'g'
         ):
     """
     Change the initial conditions parameters.
@@ -43,10 +44,13 @@ def IC_set_IC_params(
     icSimParams['periodic'] = periodic
     icSimParams['nx'] = nx
     icSimParams['ndim'] = ndim
-    if coordinate_units is None:
-        icSimParams['units'] = boxsize.units
+    if unit_l is None:
+        icSimParams['unit_l'] = boxsize.units
+        icSimParams['boxsizeToUse'] = boxsize
     else:
-        isCimParams['units'] = coordinate_units
+        icSimParams['unit_l'] = unit_l
+        icSimParams['boxsizeToUse'] = boxsize.to(unit_l)
+    icSimParams['unit_m'] = unit_m
 
     return icSimParams
 
@@ -117,7 +121,7 @@ def IC_set_run_params(
 
 
 
-def IC_uniform_coordinates(icSimParams=IC_set_IC_params()):
+def IC_uniform_coordinates(icSimParams):
     """
     Get the coordinates for a uniform particle distribution.
     nx:         number of particles in each dimension
@@ -128,7 +132,7 @@ def IC_uniform_coordinates(icSimParams=IC_set_IC_params()):
     """
 
     nx = icSimParams['nx']
-    boxsize = icSimParams['boxsize']
+    boxsize = icSimParams['boxsizeToUse']
     ndim = icSimParams['ndim']
 
     npart = nx**ndim
@@ -174,16 +178,12 @@ def IC_perturbed_coordinates(icSimParams):
     The perturbation won't exceed the half interparticle distance
     along an axis.
 
-    nx:         number of particles in each dimension
-    ndim:       number of dimensions
-    periodic:   whether we have periodic boundary conditions or not
-
     returns:
         x: np.array((nx**ndim, ndim), dtype=float) of coordinates
     """
 
     nx = icSimParams['nx']
-    boxsize = icSimParams['boxsize']
+    boxsize = icSimParams['boxsizeToUse']
     ndim = icSimParams['ndim']
     periodic = icSimParams['periodic']
     npart = nx**ndim 
@@ -243,8 +243,7 @@ def IC_sample_coordinates(icSimParams, rho_anal, rho_max = None):
 
     parameters:
         nx:         number of particles in each dimension
-        rho_anal:   function rho_anal(x, ndim). Should return a float of the analytical function rho(x)
-                    for given coordinates x
+        rho_anal:   function rho_anal(x, ndim). 
         rho_max:    peak value of the density. If none, an approximate value will be found.
         ndim:       number of dimensions
 
@@ -253,31 +252,32 @@ def IC_sample_coordinates(icSimParams, rho_anal, rho_max = None):
     """
 
     nx = icSimParams['nx']
-    boxsize = icSimParams['boxsize']
+    boxsize = icSimParams['boxsizeToUse']
     ndim = icSimParams['ndim']
     periodic = icSimParams['periodic']
     npart = nx**ndim 
 
-    x = unyt_array(np.empty((npart, 3), dtype=np.float), boxsize.units)
+    x = np.empty((npart, 3), dtype=np.float)
  
     if rho_max is None:
         # find approximate peak value of rho_max
-        nc = 200 # don't cause memory errors with too big of a grid
+        nc = 200 # don't cause memory errors with too big of a grid. Also don't worry too much about accuracy.
         xc = IC_uniform_coordinates(IC_set_IC_params(boxsize=boxsize, periodic=periodic, nx=nc, ndim=ndim))
[…]
