### 2020-07-26 c0fa221
updated docstrings
