+++ swiftsimio/initial_conditions/generate_particles.py
@@ -1,5 +1,6 @@
 """
-Generate particle initial conditions that follow some density function.
+Generate SPH initial conditions for SPH simulations iteratively for a given density function following Arth et al. 2019 (https://arxiv.org/abs/1907.11250).
+
 """
 
 #--------------------------------------------
@@ -8,11 +9,13 @@ Generate particle initial conditions that follow some density function.
 
 
 import numpy as np
+import unyt
 from unyt import unyt_array
 from math import erf
 from matplotlib import pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
 from scipy import stats
+from typing import Union
 
 from .IC_kernel import get_kernel_data
 
@@ -23,20 +26,55 @@ np.random.seed(666)
 
 
 def IC_set_IC_params(
-            boxsize     = unyt_array([1., 1., 1.], "cm"),
-            periodic    = True,
-            nx          = 100,
-            ndim        = 2,
-            unit_l      = None,
-            unit_m      = 'g'
+            boxsize: unyt_array = unyt_array([1., 1., 1.], "cm"),
+            periodic: bool = True,
+            nx : int = 100,
+            ndim : int = 2,
+            unit_l: Union[None, unyt.unit_object.Unit, str] = None,
+            unit_m: Union[unyt.unit_object.Unit, str] = 'g'
         ):
-    """
-    Change the initial conditions parameters.
+    r"""
+    Set up the simulation parameters for the initial conditions you want to
+    generate. The dict it returns is a necessary argument to call
+    ``generate_IC_for_given_density()``
 
-    TODO: docs
 
-    returns:
-        dict with these parameters.
+    Parameters
+    -------------
+
+    boxsize: unyt_array, optional
+        The box size of the simulation.
+
+    periodic: bool, optional
+        whether the simulation box is periodic or not
+
+    nx: int, optional
+        how many particles along every dimension you want your simulation
+        to contain
+
+    ndim: int, optional
+        how many dimensions you want your simulation to have
+
+    unit_l:  None or unyt.unit_object.Unit or str, optional
+        unit length for the coordinates. If None, the same unit given for
+        the boxsize will be used.
+
+    unit_m: unyt.unit_object.Unit or str, optional
+        unit mass for the particle masses
+
+
+    Returns
+    ------------
+
+    icSimParams: dict
+        dict containing the parameters stored in a way such that ``generate_IC_for_given_density()``
+        understands them.
+
+
+    Notes
+    -----------
+    
+    + The returned dict is a required argument to call ``generate_IC_for_given_density()``
     """
 
     icSimParams = {}
@@ -55,50 +93,83 @@ def IC_set_IC_params(
     return icSimParams
 
 
+
+
+
+
 def IC_set_run_params(
-            iter_max                    = 2000,
-            convergence_threshold       = 1e-3,
-            tolerance_part              = 1e-2,
-            displacement_threshold      = 1e-2,
-            delta_init                  = 0.1,
-            delta_reduction_factor      = 0.98,
-            delta_min                   = 1e-4,
-            redistribute_frequency      = 20,
-            redistribute_fraction       = 0.01,
-            no_redistribution_after     = 200,
+            iter_max: int = 2000,
+            convergence_threshold: float = 1e-3,
+            tolerance_part: float =  1e-2,
+            displacement_threshold: float =  1e-2,
+            delta_init: Union[float, None] = 0.1,
+            delta_reduction_factor: float = 1.,
+            delta_min: float =  1e-4,
+            redistribute_frequency: int = 20,
+            redistribute_fraction: float = 0.01,
+            redistribute_fraction_reduction: float = 1.0,
+            no_redistribution_after: int = 200,
             plot_at_redistribution      = True, 
[…]
