+++ swiftsimio/initial_conditions/generate_particles.py
@@ -3,9 +3,9 @@ Generate SPH initial conditions for SPH simulations iteratively for a given dens
 
 """
 
-#--------------------------------------------
+# --------------------------------------------
 # IC generation related stuff
-#--------------------------------------------
+# --------------------------------------------
 
 
 import numpy as np
@@ -26,13 +26,13 @@ np.random.seed(666)
 
 
 def IC_set_IC_params(
-            boxsize: unyt_array = unyt_array([1., 1., 1.], "cm"),
-            periodic: bool = True,
-            nx : int = 100,
-            ndim : int = 2,
-            unit_l: Union[None, unyt.unit_object.Unit, str] = None,
-            unit_m: Union[unyt.unit_object.Unit, str] = 'g'
-        ):
+    boxsize: unyt_array = unyt_array([1.0, 1.0, 1.0], "cm"),
+    periodic: bool = True,
+    nx: int = 100,
+    ndim: int = 2,
+    unit_l: Union[None, unyt.unit_object.Unit, str] = None,
+    unit_m: Union[unyt.unit_object.Unit, str] = "g",
+):
     r"""
     Set up the simulation parameters for the initial conditions you want to
     generate. The dict it returns is a necessary argument to call
@@ -78,41 +78,37 @@ def IC_set_IC_params(
     """
 
     icSimParams = {}
-    icSimParams['boxsize'] = boxsize
-    icSimParams['periodic'] = periodic
-    icSimParams['nx'] = nx
-    icSimParams['ndim'] = ndim
+    icSimParams["boxsize"] = boxsize
+    icSimParams["periodic"] = periodic
+    icSimParams["nx"] = nx
+    icSimParams["ndim"] = ndim
     if unit_l is None:
-        icSimParams['unit_l'] = boxsize.units
-        icSimParams['boxsizeToUse'] = boxsize
+        icSimParams["unit_l"] = boxsize.units
+        icSimParams["boxsizeToUse"] = boxsize
     else:
-        icSimParams['unit_l'] = unit_l
-        icSimParams['boxsizeToUse'] = boxsize.to(unit_l)
-    icSimParams['unit_m'] = unit_m
+        icSimParams["unit_l"] = unit_l
+        icSimParams["boxsizeToUse"] = boxsize.to(unit_l)
+    icSimParams["unit_m"] = unit_m
 
     return icSimParams
 
 
-
-
-
-
 def IC_set_run_params(
-            iter_max: int = 2000,
-            convergence_threshold: float = 1e-3,
-            tolerance_part: float =  1e-2,
-            displacement_threshold: float =  1e-2,
-            delta_init: Union[float, None] = 0.1,
-            delta_reduction_factor: float = 1.,
-            delta_min: float =  1e-4,
-            redistribute_frequency: int = 20,
-            redistribute_fraction: float = 0.01,
-            redistribute_fraction_reduction: float = 1.0,
-            no_redistribution_after: int = 200,
-            plot_at_redistribution      = True, 
-            kernel: str = 'cubic spline',
-            eta: float  = 1.2348
-        ):
+    iter_max: int = 2000,
+    convergence_threshold: float = 1e-3,
+    tolerance_part: float = 1e-2,
+    displacement_threshold: float = 1e-2,
+    delta_init: Union[float, None] = 0.1,
+    delta_reduction_factor: float = 1.0,
+    delta_min: float = 1e-4,
+    redistribute_frequency: int = 20,
+    redistribute_fraction: float = 0.01,
+    redistribute_fraction_reduction: float = 1.0,
[…]
