### 2020-07-26 3d9536f
added random seed parameter
