### 2020-07-26 110f1d6
added reduction of redistribution fraction
