+++ swiftsimio/initial_conditions/generate_particles.py
@@ -108,7 +108,8 @@ def IC_set_run_params(
 ):
     r"""
     Set up the runtime parameters for the initial condition generation.
-    The dict it returns is a necessary argument when calling ``generate_IC_for_given_density()``.
+    The dict it returns is a necessary argument when calling 
+    ``generate_IC_for_given_density()``.
 
     Parameters
     --------------
@@ -117,32 +118,44 @@ def IC_set_run_params(
         max numbers of iterations for generating IC conditions
 
     convergence_threshold: float, optional
-        if enough particles are displaced by distance below threshold * mean interparticle distance, stop iterating.  
+        if enough particles are displaced by distance below threshold * mean 
+        interparticle distance, stop iterating.  
     
     tolerance_part: float, optional
-        tolerance for not converged particle fraction: this fraction of particles can be displaced with distances > threshold
+        tolerance for not converged particle fraction: this fraction of 
+        particles can be displaced with distances > threshold
 
     displacement_threshold: float, optional      
-        iteration halt criterion: Don't stop until every particle is displaced by distance < threshold * mean interparticle distance
+        iteration halt criterion: Don't stop until every particle is displaced 
+        by distance < threshold * mean interparticle distance
 
     delta_init : float or None, optional
-        initial normalization constant for particle motion in units of mean interparticle distance. If None (default), delta_init will be set such that the maximal displacement found in the first iteration is normalized to 1 mean interparticle distance.
+        initial normalization constant for particle motion in units of mean 
+        interparticle distance. If None (default), delta_init will be set such 
+        that the maximal displacement found in the first iteration is normalized
+        to 1 mean interparticle distance.
 
     delta_reduction_factor: float, optional
-        multiply the normalization constant for particle motion by this factor after every iteration. In certain difficult cases this might help the generation to converge if set to < 1.
+        multiply the normalization constant for particle motion by this factor 
+        after every iteration. In certain difficult cases this might help the 
+        generation to converge if set to < 1.
     
     delta_min: float, optional
-        minimal normalization constant for particle motion in units of mean interparticle distance.
+        minimal normalization constant for particle motion in units of mean 
+        interparticle distance.
    
     redistribute_frequency: int, optional 
-        redistribute a handful of particles every ``redistribute_frequency`` iteration.
-        How many particles are redistributed is controlled with the ``redistribute_fraction`` parameter.
+        redistribute a handful of particles every ``redistribute_frequency`` 
+        iteration. How many particles are redistributed is controlled with the 
+        ``redistribute_fraction`` parameter.
 
     redistribute_fraction: float, optional
         fraction of particles to be redistributed when doing so. 
 
     redistribute_fraction_reduction: float, optional
-        multiply the ``redistribute_fraction`` parameter by this factor every time particles are being redistributed. In certain difficult cases this might help the generation to converge if set to < 1.
+        multiply the ``redistribute_fraction`` parameter by this factor every 
+        time particles are being redistributed. In certain difficult cases this
+        might help the generation to converge if set to < 1.
     
     no_redistribution_after: int, optional     
         don't redistribute particles after this iteration.
@@ -159,13 +172,15 @@ def IC_set_run_params(
     ------------
 
     icRunParams: dict
-        dict containing the parameters stored in a way such that ``generate_IC_for_given_density()`` understands them.
[…]
