### 2020-07-26 7dbd736
added niter to stats dict
