### 2020-07-26 4627862
delta_init can be generated automatically now
