### 2020-07-26 408353e
added 3D case plotting
