+++ swiftsimio/initial_conditions/generate_particles.py
@@ -576,15 +576,15 @@ def generate_IC_for_given_density(
     # use unitless arrays from this point on
     x_nounit = x.value
 
-    if icRunParams["PLOT_AT_REDISTRIBUTION"]:
-        # drop a first plot
-        tree = KDTree(x_nounit[:, :ndim], boxsize=boxsizeForTree)
-        for p in range(npart):
-            dist, neighs = tree.query(x_nounit[p, :ndim], k=Nngb_int)
-            h[p] = dist[-1] / kernel_gamma
-            W = kernel_func(dist, dist[-1])
-            rho[p] = (W * m[neighs]).sum()
-        _plot_current_situation(True, 0, x_nounit, rho, rho_anal, icSimParams)
+    #  if icRunParams["PLOT_AT_REDISTRIBUTION"]:
+    #      # drop a first plot
+    #      tree = KDTree(x_nounit[:, :ndim], boxsize=boxsizeForTree)
+    #      for p in range(npart):
+    #          dist, neighs = tree.query(x_nounit[p, :ndim], k=Nngb_int)
+    #          h[p] = dist[-1] / kernel_gamma
+    #          W = kernel_func(dist, dist[-1])
+    #          rho[p] = (W * m[neighs]).sum()
+    #      _plot_current_situation(True, 0, x_nounit, rho, rho_anal, icSimParams)
 
     # start iteration loop
     iteration = 0
@@ -941,43 +941,41 @@ def _plot_current_situation(
     + For debugging/checking purposes only, not meant to be called.
     """
     from matplotlib import pyplot as plt
-    from mpl_toolkits.mplot3d import Axes3D
     from scipy import stats
 
     boxsize = icSimParams["boxsizeToUse"]
     ndim = icSimParams["ndim"]
+    nx = icSimParams["nx"]
 
     if x.shape[0] > 5000:
         marker = ","
     else:
         marker = "."
 
+    # dict to generate coordinates to compute analytical solution
+    tempparams = IC_set_IC_params(
+        boxsize=icSimParams["boxsize"],
+        periodic=icSimParams["periodic"],
+        nx=200,
+        ndim=1,
+        unit_l=icSimParams["unit_l"],
+    )
+
     if ndim == 1:
+
         fig = plt.figure(figsize=(6, 6), dpi=200)
         ax1 = fig.add_subplot(111)
         ax1.scatter(x[:, 0], rho, s=1, c="k", label="IC")
 
-        r = np.zeros((100, 3), dtype=np.float)
-        r[:, 0] = np.linspace(0, boxsize[0], 100)
-        ax1.plot(r[:, 0], rho_anal(r, ndim), label="analytical")
+        XA = IC_uniform_coordinates(tempparams)
+        ax1.plot(XA[:, 0], rho_anal(XA, ndim), label="analytical")
         ax1.set_xlabel("x")
         ax1.set_ylabel("rho")
         ax1.legend()
 
     elif ndim == 2:
-        fig = plt.figure(figsize=(18, 6), dpi=200)
 
-        # get rho map
-        #  X = np.linspace(0, 1, 20)
-        #  Y = np.linspace(0, 1, 20)
-        #  X, Y = np.meshgrid(X, Y)
-        #  xmap = np.stack((X.ravel(), Y.ravel()), axis=1)
-        #  rhomap = rho_anal(xmap).reshape(20,20)
-        #
-        #  ax1 = fig.add_subplot(1,1,1, projection='3d')
-        #  im1 = ax1.plot_surface(X, Y, rhomap, color='red', alpha=0.2)
[…]
