### 2020-07-26 7795323
boxsizeToUse is now numpy array, not unyt
