### 2020-07-27 6f674fe
added intermediate file dump
