+++ swiftsimio/initial_conditions/generate_particles.py
@@ -28,6 +28,8 @@ def IC_set_IC_params(
     ndim: int = 2,
     unit_l: Union[None, unyt.unit_object.Unit, str] = None,
     unit_m: Union[unyt.unit_object.Unit, str] = "g",
+    kernel: str = "cubic spline",
+    eta: float = 1.2348,
 ):
     r"""
     Set up the simulation parameters for the initial conditions you want to
@@ -58,6 +60,13 @@ def IC_set_IC_params(
     unit_m: unyt.unit_object.Unit or str, optional
         unit mass for the particle masses
 
+    kernel: str {'cubic spline',}, optional
+        which kernel to use
+
+    eta: float, optional
+        resolution eta, which defines the number of neighbours used
+        independently of dimensionality
+
 
     Returns
     ------------
@@ -85,6 +94,8 @@ def IC_set_IC_params(
         icSimParams["unit_l"] = unit_l
         icSimParams["boxsizeToUse"] = boxsize.to(unit_l).value
     icSimParams["unit_m"] = unit_m
+    icSimParams["kernel"] = kernel
+    icSimParams["eta"] = eta
 
     return icSimParams
 
@@ -92,18 +103,16 @@ def IC_set_IC_params(
 def IC_set_run_params(
     iter_max: int = 2000,
     convergence_threshold: float = 1e-3,
-    tolerance_part: float = 1e-2,
-    displacement_threshold: float = 1e-2,
+    tolerance_part: float = 1e-3,
+    displacement_threshold: float = 5e-3,
     delta_init: Union[float, None] = None,
     delta_reduction_factor: float = 1.0,
-    delta_min: float = 1e-4,
+    delta_min: float = 1e-6,
     redistribute_frequency: int = 20,
     redistribute_fraction: float = 0.01,
     redistribute_fraction_reduction: float = 1.0,
     no_redistribution_after: int = 200,
-    plot_at_redistribution=True,
-    kernel: str = "cubic spline",
-    eta: float = 1.2348,
+    intermediate_dump_frequency: int = 50,
     random_seed: int = 666,
 ):
     r"""
@@ -160,8 +169,9 @@ def IC_set_run_params(
     no_redistribution_after: int, optional     
         don't redistribute particles after this iteration.
  
-    plot_at_redistribution: bool, optional
-        create and store a plot of the current situation before redistributing?
+    intermediate_dump_frequency: int, optional
+        frequency of dumps of the current state of the iteration. If set to zero,
+        no intermediate results will be stored.
 
     random_seed: int, optional
         set a specific random seed
@@ -196,9 +206,7 @@ def IC_set_run_params(
     icRunParams["REDISTRIBUTE_FRACTION"] = redistribute_fraction
     icRunParams["NO_REDISTRIBUTION_AFTER"] = no_redistribution_after
     icRunParams["REDISTRIBUTE_FRACTION_REDUCTION"] = redistribute_fraction_reduction
-    icRunParams["PLOT_AT_REDISTRIBUTION"] = plot_at_redistribution
-    icRunParams["KERNEL"] = kernel
-    icRunParams["ETA"] = eta
+    icRunParams["DUMPFREQ"] = intermediate_dump_frequency
 
     global seed_set
     if not seed_set:
@@ -537,15 +545,15 @@ def generate_IC_for_given_density(
         boxsizeForTree = None
 
     # kernel data
-    kernel_func, _, kernel_gamma = get_kernel_data(icRunParams["KERNEL"], ndim)
+    kernel_func, _, kernel_gamma = get_kernel_data(icSimParams["kernel"], ndim)
 
     # expected number of neighbours
     if ndim == 1:
-        Nngb = 2 * kernel_gamma * icRunParams["ETA"]
[…]
