### 2020-07-27 fab1396
adapted kernels following Josh's suggestions
