### 2020-07-27 3242b6b
variable name conventions
