+++ swiftsimio/initial_conditions/generate_particles.py
@@ -71,7 +71,7 @@ def IC_set_IC_params(
     Returns
     ------------
 
-    icSimParams: dict
+    ic_sim_params: dict
         dict containing the parameters stored in a way such that ``generate_IC_for_given_density()``
         understands them.
 
@@ -82,22 +82,22 @@ def IC_set_IC_params(
     + The returned dict is a required argument to call ``generate_IC_for_given_density()``
     """
 
-    icSimParams = {}
-    icSimParams["boxsize"] = boxsize
-    icSimParams["periodic"] = periodic
-    icSimParams["nx"] = nx
-    icSimParams["ndim"] = ndim
+    ic_sim_params = {}
+    ic_sim_params["boxsize"] = boxsize
+    ic_sim_params["periodic"] = periodic
+    ic_sim_params["nx"] = nx
+    ic_sim_params["ndim"] = ndim
     if unit_l is None:
-        icSimParams["unit_l"] = boxsize.units
-        icSimParams["boxsizeToUse"] = boxsize.value  # use only np.array
+        ic_sim_params["unit_l"] = boxsize.units
+        ic_sim_params["boxsizeToUse"] = boxsize.value  # use only np.array
     else:
-        icSimParams["unit_l"] = unit_l
-        icSimParams["boxsizeToUse"] = boxsize.to(unit_l).value
-    icSimParams["unit_m"] = unit_m
-    icSimParams["kernel"] = kernel
-    icSimParams["eta"] = eta
+        ic_sim_params["unit_l"] = unit_l
+        ic_sim_params["boxsizeToUse"] = boxsize.to(unit_l).value
+    ic_sim_params["unit_m"] = unit_m
+    ic_sim_params["kernel"] = kernel
+    ic_sim_params["eta"] = eta
 
-    return icSimParams
+    return ic_sim_params
 
 
 def IC_set_run_params(
@@ -181,7 +181,7 @@ def IC_set_run_params(
     Returns
     ------------
 
-    icRunParams: dict
+    ic_run_params: dict
         dict containing the parameters stored in a way such that 
         ``generate_IC_for_given_density()`` understands them.
 
@@ -193,30 +193,30 @@ def IC_set_run_params(
         ``generate_IC_for_given_density()``
     """
 
-    icRunParams = {}
+    ic_run_params = {}
 
-    icRunParams["ITER_MAX"] = iter_max
-    icRunParams["CONVERGE_THRESH"] = convergence_threshold
-    icRunParams["TOLERANCE_PART"] = tolerance_part
-    icRunParams["DISPL_THRESH"] = displacement_threshold
-    icRunParams["DELTA_INIT"] = delta_init
-    icRunParams["DELTA_REDUCT"] = delta_reduction_factor
-    icRunParams["DELTA_MIN"] = delta_min
-    icRunParams["REDIST_FREQ"] = redistribute_frequency
-    icRunParams["REDIST_FRAC"] = redistribute_fraction
-    icRunParams["REDIST_STOP"] = no_redistribution_after
-    icRunParams["REDIST_REDUCT"] = redistribute_fraction_reduction
-    icRunParams["DUMPFREQ"] = intermediate_dump_frequency
+    ic_run_params["ITER_MAX"] = iter_max
+    ic_run_params["CONVERGE_THRESH"] = convergence_threshold
+    ic_run_params["TOLERANCE_PART"] = tolerance_part
+    ic_run_params["DISPL_THRESH"] = displacement_threshold
+    ic_run_params["DELTA_INIT"] = delta_init
+    ic_run_params["DELTA_REDUCT"] = delta_reduction_factor
+    ic_run_params["DELTA_MIN"] = delta_min
[…]
