### 2020-07-27 63dce5a
various small fixes
