### 2020-07-27 b6f5ad8
small fixes & parameters are objects now everywhere. Code works.
