+++ swiftsimio/initial_conditions/generate_particles.py
@@ -22,202 +22,247 @@ from swiftsimio import Writer
 seed_set = False
 
 
-def IC_set_IC_params(
-    boxsize: unyt_array = unyt_array([1.0, 1.0, 1.0], "cm"),
-    periodic: bool = True,
-    nx: int = 100,
-    ndim: int = 2,
-    unit_l: Union[None, unyt.unit_object.Unit, str] = None,
-    unit_m: Union[unyt.unit_object.Unit, str] = "g",
-    kernel: str = "cubic spline",
-    eta: float = 1.2348,
-):
+class ic_sim_params(object):
     r"""
-    Set up the simulation parameters for the initial conditions you want to
-    generate. The dict it returns is a necessary argument to call
-    ``generate_IC_for_given_density()``
-
-
-    Parameters
-    -------------
-
-    boxsize: unyt_array, optional
-        The box size of the simulation.
-
-    periodic: bool, optional
-        whether the simulation box is periodic or not
-
-    nx: int, optional
-        how many particles along every dimension you want your simulation
-        to contain
-
-    ndim: int, optional
-        how many dimensions you want your simulation to have
-
-    unit_l:  None or unyt.unit_object.Unit or str, optional
-        unit length for the coordinates. If None, the same unit given for
-        the boxsize will be used.
-
-    unit_m: unyt.unit_object.Unit or str, optional
-        unit mass for the particle masses
-
-    kernel: str {'cubic spline',}, optional
-        which kernel to use
-
-    eta: float, optional
-        resolution eta, which defines the number of neighbours used
-        independently of dimensionality
-
-
-    Returns
-    ------------
-
-    ic_sim_params: dict
-        dict containing the parameters stored in a way such that ``generate_IC_for_given_density()``
-        understands them.
-
-
-    Notes
-    -----------
-    
-    + The returned dict is a required argument to call ``generate_IC_for_given_density()``
+    object containing simulation parameters
     """
+    boxsize: unyt_array = unyt_array([1.0, 1.0, 1.0], "cm")
+    periodic: bool = True
+    nx: int = 100
+    ndim: int = 2
+    unit_l: Union[None, unyt.unit_object.Unit, str] = None
+    unit_m: Union[unyt.unit_object.Unit, str] = "g"
+    kernel: str = "cubic spline"
+    eta: float = 1.2348
 
-    ic_sim_params = {}
-    ic_sim_params["boxsize"] = boxsize
-    ic_sim_params["periodic"] = periodic
-    ic_sim_params["nx"] = nx
-    ic_sim_params["ndim"] = ndim
-    if unit_l is None:
-        ic_sim_params["unit_l"] = boxsize.units
-        ic_sim_params["boxsizeToUse"] = boxsize.value  # use only np.array
-    else:
-        ic_sim_params["unit_l"] = unit_l
-        ic_sim_params["boxsizeToUse"] = boxsize.to(unit_l).value
-    ic_sim_params["unit_m"] = unit_m
-    ic_sim_params["kernel"] = kernel
-    ic_sim_params["eta"] = eta
-
-    return ic_sim_params
-
-
-def IC_set_run_params(
-    iter_max: int = 2000,
-    convergence_threshold: float = 1e-4,
-    tolerance_part: float = 1e-3,
-    displacement_threshold: float = 1e-3,
-    delta_init: Union[float, None] = None,
-    delta_reduction_factor: float = 1.0,
-    delta_min: float = 1e-6,
-    redistribute_frequency: int = 20,
-    redistribute_fraction: float = 0.01,
-    redistribute_fraction_reduction: float = 1.0,
-    no_redistribution_after: int = 200,
-    intermediate_dump_frequency: int = 50,
-    random_seed: int = 666,
-):
-    r"""
[…]
