### 2020-07-27 33688a2
moved plotting to its own file
