+++ swiftsimio/initial_conditions/IC_plotting.py
@@ -1,0 +1,283 @@
+import numpy as np
+import unyt
+from .generate_particles import ic_sim_params
+
+
+def IC_plot_current_situation(
+    save: bool,
+    iteration: int,
+    x: np.ndarray,
+    rho: np.ndarray,
+    rho_anal: callable,
+    ic_sim_params: ic_sim_params,
+):
+    r"""
+    Create a plot of what things look like now. In particular, scatter the
+    particle positions and show what the densities currently look like.
+
+    Parameters
+    ----------------
+
+    save: bool
+        Whether to save to file (if True), or just show a plot.
+
+    iteration: int 
+        Current iteration number.
+
+    x: np.ndarray 
+        particle positions. Must be numpy array.
+
+    rho: np.ndarray
+        numpy array of SPH densities at particle positions
+
+    rho_anal: callable
+        The density function that is to be reproduced in the initial conditions.
+        It must take two positional arguments:
+        
+        - ``x``: np.ndarray of 3D particle coordinates (even if your initial 
+          conditions have lower dimensionality)
+        - ``ndim``: integer, number of dimensions that your simulations is to 
+          have
+    
+    ic_sim_params: ic_sim_params
+        an ``ic_sim_params`` instance containing simulation parameters
+
+    Note
+    ---------------
+
+    + For debugging/checking purposes only, not meant to be called.
+    """
+    from matplotlib import pyplot as plt
+    from scipy import stats
+
+    # do this here to avoid circular imports
+    from .generate_particles import IC_uniform_coordinates
+
+    boxsize = ic_sim_params.boxsize_to_use
+    ndim = ic_sim_params.ndim
+    nx = ic_sim_params.nx
+
+    if x.shape[0] > 5000:
+        marker = ","
+    else:
+        marker = "."
+
+    # dict to generate coordinates to compute analytical solution
+    tempparams = ic_sim_params.copy()
+    tempparams.ndim = 1
+    tempparams.nx = 200
+
+    if ndim == 1:
+
+        fig = plt.figure(figsize=(6, 6), dpi=200)
+        ax1 = fig.add_subplot(111)
+        ax1.scatter(x[:, 0], rho, s=1, c="k", label="IC")
+
+        XA = IC_uniform_coordinates(tempparams).value
+        ax1.plot(XA[:, 0], rho_anal(XA, ndim), label="analytical")
+        ax1.set_xlabel("x")
+        ax1.set_ylabel("rho")
+        ax1.legend()
+
+    elif ndim == 2:
+
+        fig = plt.figure(figsize=(18, 6), dpi=200)
+
+        ax1 = fig.add_subplot(131, aspect="equal")
+        ax2 = fig.add_subplot(132,)
+        ax3 = fig.add_subplot(133,)
+
+        # x - y scatterplot
+
+        ax1.scatter(x[:, 0], x[:, 1], s=1, alpha=0.5, c="k", marker=marker)
+        ax1.set_xlabel("x")
+        ax1.set_ylabel("y")
+
+        # x - rho plot
+        rho_mean, edges, _ = stats.binned_statistic(
+            x[:, 0], rho, statistic="mean", bins=min(nx, 50)
+        )
+        rho2mean, edges, _ = stats.binned_statistic(
[…]
