+++ swiftsimio/initial_conditions/generate_particles.py
@@ -276,7 +276,7 @@ def IC_uniform_coordinates(ic_sim_params: ic_sim_params):
     Parameters
     ------------------
 
-    ic_sim_params: swiftsimio.initional_conditions.ic_sim_params
+    ic_sim_params: ic_sim_params
         an ``ic_sim_params`` instance containing simulation parameters
 
 
@@ -286,8 +286,7 @@ def IC_uniform_coordinates(ic_sim_params: ic_sim_params):
 
     x: unyt_array 
         unyt_array(shape=(npart, 3), dtype=float) of coordinates, where 
-        ``npart = nx ** ndim``, both of which are set in the ``ic_sim_params``
-        dict
+        ``npart = nx ** ndim``, both of which are set in ``ic_sim_params``
     """
 
     nx = ic_sim_params.nx
@@ -445,8 +444,7 @@ def IC_sample_coordinates(
 
     x: unyt_array 
         unyt_array(shape=(npart, 3), dtype=float) of coordinates, where 
-        ``npart = nx ** ndim``, both of which are set in the ``ic_sim_params`` 
-        dict
+        ``npart = nx ** ndim``, both of which are set in ``ic_sim_params`` 
     """
 
     nx = ic_sim_params.nx
@@ -548,6 +546,9 @@ def generate_IC_for_given_density(
           experienced in units of mean interparticle distance
         - ``stats['niter']``: Number of iterations performed
     """
+    # this import is only temporary to avoid circular imports.
+    # it will be removed before the merge.
+    from .IC_plotting import IC_plot_current_situation
 
     # safety checks first
 
@@ -655,10 +656,10 @@ def generate_IC_for_given_density(
             for i, n in enumerate(neighs):
                 W = kernel_func(dist[i], dist[-1])
                 rho[p] += W * m[n]
-        _IC_write_intermediate_output(0, x, m, rho, h, ic_sim_params, ic_run_params)
+        IC_write_intermediate_output(0, x, m, rho, h, ic_sim_params, ic_run_params)
         # drop a first plot
         # TODO: remove the plotting
-        _IC_plot_current_situation(True, 0, x_nounit, rho, rho_anal, ic_sim_params)
+        IC_plot_current_situation(True, 0, x_nounit, rho, rho_anal, ic_sim_params)
 
     # start iteration loop
     iteration = 0
@@ -690,11 +691,11 @@ def generate_IC_for_given_density(
                     rho[p] += W * m[n]
 
             if dump_now:
-                _IC_write_intermediate_output(
+                IC_write_intermediate_output(
                     iteration, x_nounit, m, rho, h, ic_sim_params, ic_run_params
                 )
                 # TODO: remove the plotting
-                _IC_plot_current_situation(
+                IC_plot_current_situation(
                     True, iteration, x_nounit, rho, rho_anal, ic_sim_params
                 )
 
@@ -969,7 +970,7 @@ def redistribute_particles(
     return x, indices[touched]
 
 
-def _IC_write_intermediate_output(
+def IC_write_intermediate_output(
     iteration: int,
     x: np.ndarray,
     m: np.ndarray,
@@ -1039,283 +1040,5 @@ def _IC_write_intermediate_output(
     # If IDs are not present, this automatically generates
[…]
