### 2020-07-27 e4d3abe
added check for boxsize being unyt_array
