### 2020-07-28 a6e384f
removed plotting, reduced printing for mahsa
