### 2020-07-28 7789ce9
added script for mahsa and small bugfix. script should work now.
