### 2020-07-29 5e5b377
better version of OOP
