+++ swiftsimio/initial_conditions/generate_particles.py
@@ -19,25 +19,10 @@ from swiftsimio.optional_packages import KDTree, TREE_AVAILABLE
 from swiftsimio import Writer
 
 
-seed_set = False
-
-
-class ParticleGenerator(object):
-
-    # simulation parameters
-    rhofunc: callable
-    rhomax: Union[float, None] = None
-    boxsize: unyt_array = unyt_array([1.0, 1.0, 1.0], "cm")
-    periodic: bool = True
-    nx: int = 100
-    ndim: int = 2
-    x: Union[unyt_array, None] = None
-    m: Union[unyt_array, None] = None
-    unit_l: Union[None, unyt.unit_object.Unit, str] = None
-    unit_m: Union[unyt.unit_object.Unit, str] = "g"
-    kernel: str = "cubic spline"
-    eta: float = 1.2348
-
+class RunParams(object):
+    r"""
+    TODO: dox
+    """
     # iteration parameters
     iter_max: int = 2000
     converge_thresh: float = 1e-4
@@ -49,167 +34,216 @@ class ParticleGenerator(object):
     redist_freq: int = 20
     redist_frac: float = 0.01
     redist_reduct: float = 1.0
-    no_redistribution_after: int = 200
+    redist_stop: int = 200
     dump_basename: str = "IC-generation-iteration-"
     dumpfreq: int = 50
 
+    def __init__(self):
+        # always set one random seed.
+        self.set_random_seed(666)
+        return
+
+    def set_random_seed(self, seed: int = 666):
+        r"""
+        TODO: dox
+        """
+        np.random.seed(seed)
+        return
+
+
+class IterData(object):
+    r"""
+    TODO: dox
+    contains data relevant during the iteration
+    """
+
+    mean_interparticle_distance: float = 1.0
+    delta_r_norm: float = 1.0
+    delta_r_norm_min: float = 1e-2
+    compute_delta_r_norm: bool = True
+    Nngb: float = 20.0
+    Nngb_int: int = 20
+    boxsize_for_tree: np.ndarray = np.ones(3, dtype=float)
+
+    def __init__(self):
+        return
+
+
+class ParticleGenerator(object):
+    r"""
+
+    TODO: dox
+
+    Set up the simulation parameters for the initial conditions you want to
+    generate. The dict it returns is a necessary argument to call
+    ``generate_IC_for_given_density()``
+
+
+    Parameters
+    -------------
+
+    rho: callable
+        The density function that is to be reproduced in the initial conditions.
+        It must take two positional arguments:
+        
+        - ``x``: np.ndarray of 3D particle coordinates (even if your initial 
+          conditions have lower dimensionality)
+        - ``ndim``: integer, number of dimensions that your simulations is to 
+          have
+
+    rho_max: float or None, optional
+        The maximal density within the simulation box. If ``None``, an 
+        approximate value will be determined if the rejection sampling to obtain
+        an initial particle configuration guess is used.
+
+    boxsize: unyt_array, optional
+        The box size of the simulation.
+
+    periodic: bool, optional
+        whether the simulation box is periodic or not
+
+    nx: int, optional
+        how many particles along every dimension you want your simulation
+        to contain
+
+    ndim: int, optional
+        how many dimensions you want your simulation to have
+
+    x: unyt_array or None, optional
+        Initial guess for coordinates of particles. If ``None``, the initial 
+        guess will be generated by rejection sampling the density function 
+        ``rhofunc``
+
+    m: unyt_array or None, optional
+        ``unyt_array`` of particle masses. If ``None``, an array will be created
+        such that the total mass in the simulation box given the analytical 
+        density is reproduced, and all particles will have equal masses.
+
[…]
