### 2020-07-29 d79de0f
cleanup and improvements
