+++ swiftsimio/initial_conditions/generate_particles.py
@@ -2,12 +2,6 @@
 2019 (https://arxiv.org/abs/1907.11250).
 
 """
-
-# --------------------------------------------
-# IC generation related stuff
-# --------------------------------------------
-
-
 import numpy as np
 import unyt
 from unyt import unyt_array
@@ -22,6 +16,67 @@ from swiftsimio import Writer
 class RunParams(object):
     r"""
     TODO: dox
+
+
+    Attributes
+    ---------------
+    iter_max: int, optional
+        max numbers of iterations for generating IC conditions
+
+    converge_thresh: float, optional
+        if enough particles are displaced by distance below threshold * mean
+        interparticle distance, stop iterating.
+
+    tolerance_part: float, optional
+        tolerance for not converged particle fraction: this fraction of
+        particles can be displaced with distances > threshold
+
+    displacement_threshold: float, optional
+        iteration halt criterion: Don't stop until every particle is displaced
+        by distance < threshold * mean interparticle distance
+
+    delta_init: float or None, optional
+        initial normalization constant for particle motion in units of mean
+        interparticle distance. If ``None`` (default), delta_init will be set
+        such that the maximal displacement found in the first iteration is
+        normalized to 1 mean interparticle distance.
+
+    delta_reduction_factor: float, optional
+        multiply the normalization constant for particle motion by this factor
+        after every iteration. In certain difficult cases this might help the
+        generation to converge if set to < 1.
+
+    delta_min: float, optional
+        minimal normalization constant for particle motion in units of mean
+        interparticle distance.
+
+    redistribute_frequency: int, optional
+        redistribute a handful of particles every ``redistribute_frequency``
+        iteration. How many particles are redistributed is controlled with the
+        ``redistribute_fraction`` parameter.
+
+    redistribute_fraction: float, optional
+        fraction of particles to be redistributed when doing so.
+
+    redistribute_fraction_reduction: float, optional
+        multiply the ``redistribute_fraction`` parameter by this factor every
+        time particles are being redistributed. In certain difficult cases this
+        might help the generation to converge if set to < 1.
+
+    no_redistribution_after: int, optional
+        don't redistribute particles after this iteration.
+
+    intermediate_dump_frequency: int, optional
+        frequency of dumps of the current state of the iteration. If set to zero,
+        no intermediate results will be stored.
+
+    Tdump_basename: str
+        Basename for intermediate dumps. The filename will be constructed as
+        ``basename + <5 digit zero padded iteration number> + .hdf5``
+
+    random_seed: int, optional
+        set a specific random seed
+
     """
     # iteration parameters
     iter_max: int = 2000
@@ -109,16 +164,6 @@ class ParticleGenerator(object):
     ndim: int, optional
         how many dimensions you want your simulation to have
 
-    x: unyt_array or None, optional
-        Initial guess for coordinates of particles. If ``None``, the initial 
-        guess will be generated by rejection sampling the density function 
-        ``rhofunc``
-
-    m: unyt_array or None, optional
-        ``unyt_array`` of particle masses. If ``None``, an array will be created
-        such that the total mass in the simulation box given the analytical 
-        density is reproduced, and all particles will have equal masses.
-
     unit_l:  None or unyt.unit_object.Unit or str, optional
         unit length for the coordinates. If None, the same unit given for
         the boxsize will be used.
@@ -132,63 +177,6 @@ class ParticleGenerator(object):
     eta: float, optional
[…]
