### 2020-07-29 bc723f7
sped up plotting
