+++ swiftsimio/initial_conditions/generate_particles.py
@@ -146,30 +146,26 @@ class ParticleGenerator(object):
         - ``ndim``: integer, number of dimensions that your simulations is to 
           have
 
-    rho_max: float or None, optional
-        The maximal density within the simulation box. If ``None``, an 
-        approximate value will be determined if the rejection sampling to obtain
-        an initial particle configuration guess is used.
-
-    boxsize: unyt_array, optional
+    boxsize: unyt_array
         The box size of the simulation.
+
+    nx: int
+        how many particles along every dimension you want your simulation
+        to contain
+
+    ndim: int
+        how many dimensions you want your simulation to have
+
+    unitsys: unyt.unit_systems.UnitSystem
+        a unit system that contains the units you want your ICs to have
 
     periodic: bool, optional
         whether the simulation box is periodic or not
 
-    nx: int, optional
-        how many particles along every dimension you want your simulation
-        to contain
-
-    ndim: int, optional
-        how many dimensions you want your simulation to have
-
-    unit_l:  None or unyt.unit_object.Unit or str, optional
-        unit length for the coordinates. If None, the same unit given for
-        the boxsize will be used.
-
-    unit_m: unyt.unit_object.Unit or str, optional
-        unit mass for the particle masses
+    rho_max: float or None, optional
+        The maximal density within the simulation box. If ``None``, an 
+        approximate value will be determined if the rejection sampling to obtain
+        an initial particle configuration guess is used.
 
     kernel: str {'cubic spline',}, optional
         which kernel to use
@@ -188,15 +184,14 @@ class ParticleGenerator(object):
 
     # simulation parameters
     rhofunc: callable
-    rhomax: Union[float, None] = None
-    boxsize: unyt_array = unyt_array([1.0, 1.0, 1.0], "cm")
+    boxsize: unyt_array
+    nx: int
+    ndim: int
     periodic: bool = True
-    nx: int = 100
-    ndim: int = 2
-    unit_l: Union[None, unyt.unit_object.Unit, str] = None
-    unit_m: Union[unyt.unit_object.Unit, str] = "g"
+    unitsys: unyt.unit_systems.UnitSystem
     kernel: str = "cubic spline"
     eta: float = 1.2348
+    rhomax: Union[float, None] = None
 
     runparams: RunParams = RunParams()
     iterparams: IterData = IterData()
@@ -220,13 +215,12 @@ class ParticleGenerator(object):
     def __init__(
         self,
         rho: callable,
-        rho_max: Union[float, None] = None,
-        boxsize: unyt_array = unyt_array([1.0, 1.0, 1.0], "cm"),
+        boxsize: unyt_array,
+        unitsys: unyt.unit_systems.UnitSystem,
+        nx: int,
+        ndim: int,
         periodic: bool = True,
-        nx: int = 100,
-        ndim: int = 2,
-        unit_l: Union[None, unyt.unit_object.Unit, str] = None,
-        unit_m: Union[unyt.unit_object.Unit, str] = "g",
+        rho_max: Union[float, None] = None,
         kernel: str = "cubic spline",
         eta: float = 1.2348,
     ):
@@ -234,24 +228,29 @@ class ParticleGenerator(object):
         if not isinstance(boxsize, unyt.unyt_array):
             raise TypeError("boxsize needs to be a unyt array.")
 
+        if not isinstance(unitsys, unyt.unit_systems.UnitSystem):
+            raise TypeError("unitsys needs to be a unyt UnitSystem.")
+
+        if not isinstance(nx, int):
+            raise TypeError("nx needs to be an integer")
[…]
