### 2020-07-29 c8e4fb0
small fixes, added warning if max displ > 5
