+++ swiftsimio/initial_conditions/generate_particles.py
@@ -4,9 +4,9 @@
 """
 import numpy as np
 import unyt
-from unyt import unyt_array
 from math import erf
 from typing import Union
+import warnings
 
 from .IC_kernel import get_kernel_data
 from swiftsimio.optional_packages import KDTree, TREE_AVAILABLE
@@ -146,7 +146,7 @@ class ParticleGenerator(object):
         - ``ndim``: integer, number of dimensions that your simulations is to 
           have
 
-    boxsize: unyt_array
+    boxsize: unyt.unyt_array
         The box size of the simulation.
 
     nx: int
@@ -184,7 +184,7 @@ class ParticleGenerator(object):
 
     # simulation parameters
     rhofunc: callable
-    boxsize: unyt_array
+    boxsize: unyt.unyt_array
     nx: int
     ndim: int
     periodic: bool = True
@@ -203,10 +203,10 @@ class ParticleGenerator(object):
     _set_up = False
 
     # unyt/result arrays
-    coordinates: Union[unyt_array, None] = None
-    masses: Union[unyt_array, None] = None
-    smoothing_lengths: Union[unyt_array, None] = None
-    densities: Union[unyt_array, None] = None
+    coordinates: Union[unyt.unyt_array, None] = None
+    masses: Union[unyt.unyt_array, None] = None
+    smoothing_lengths: Union[unyt.unyt_array, None] = None
+    densities: Union[unyt.unyt_array, None] = None
 
     # unitless arrays to work with
     x: np.array
@@ -215,7 +215,7 @@ class ParticleGenerator(object):
     def __init__(
         self,
         rho: callable,
-        boxsize: unyt_array,
+        boxsize: unyt.unyt_array,
         unitsys: unyt.unit_systems.UnitSystem,
         nx: int,
         ndim: int,
@@ -257,19 +257,19 @@ class ParticleGenerator(object):
     def initial_setup(
         self,
         method: str = "rejection",
-        x: Union[unyt_array, None] = None,
-        m: Union[unyt_array, None] = None,
+        x: Union[unyt.unyt_array, None] = None,
+        m: Union[unyt.unyt_array, None] = None,
         max_displ: float = 0.4,
     ):
         r"""
         TODO: dox
-        x: unyt_array or None, optional
+        x: unyt.unyt_array or None, optional
             Initial guess for coordinates of particles. If ``None``, the initial
             guess will be generated by rejection sampling the density function
             ``rhofunc``
 
-        m: unyt_array or None, optional
-            ``unyt_array`` of particle masses. If ``None``, an array will be created
+        m: unyt.unyt_array or None, optional
+            ``unyt.unyt_array`` of particle masses. If ``None``, an array will be created
             such that the total mass in the simulation box given the analytical
             density is reproduced, and all particles will have equal masses.
 
@@ -310,7 +310,7 @@ class ParticleGenerator(object):
             else:
                 raise ValueError("Unknown coordinate generation method:", method)
         else:
-            if not isinstance(x, unyt_array):
+            if not isinstance(x, unyt.unyt_array):
                 raise TypeError("x must be an unyt array")
             # TODO: compare given x array units with unit system first
             self.coordinates = x
@@ -334,12 +334,12 @@ class ParticleGenerator(object):
             mtot = rhotot * area  # go from density to mass
 
             self.m = np.ones(npart, dtype=np.float) * mtot / npart
-            self.masses = unyt_array(self.m, self.unitsys["mass"])
+            self.masses = unyt.unyt_array(self.m, self.unitsys["mass"])
             print("Assigning particle mass: {0:.3e}".format(mtot / npart))
 
         else:
[…]
