### 2020-07-29 14651db
small fixes; added stats object
