+++ swiftsimio/initial_conditions/generate_particles.py
@@ -124,6 +124,41 @@ class IterData(object):
         return
 
 
+class IterStats(object):
+    r"""
+    used to store stats of the iteration.
+    """
+
+    max_displacement: np.ndarray
+    min_displacement: np.ndarray
+    avg_displacement: np.ndarray
+    niter: int
+
+    def __init__(self, iter_max):
+        self.max_displacement = np.zeros(iter_max, dtype=np.float)
+        self.min_displacement = np.zeros(iter_max, dtype=np.float)
+        self.avg_displacement = np.zeros(iter_max, dtype=np.float)
+        self.niter = 0
+        return
+
+    def add_stat(
+        self,
+        iteration,
+        min_displacement: float,
+        max_displacement: float,
+        avg_displacement: float,
+    ):
+        r"""
+        Store new stats at the appropriate place.
+        """
+
+        self.max_displacement[iteration] = max_displacement
+        self.min_displacement[iteration] = min_displacement
+        self.avg_displacement[iteration] = avg_displacement
+        self.niter = iteration
+        return
+
+
 class ParticleGenerator(object):
     r"""
 
@@ -195,6 +230,7 @@ class ParticleGenerator(object):
 
     runparams: RunParams = RunParams()
     iterparams: IterData = IterData()
+    stats: IterStats
 
     # derived variables
     npart = 0
@@ -317,7 +353,7 @@ class ParticleGenerator(object):
 
         #  generate masses if necessary
         if m is None:
-            nc = 1000 - 400 * (ndim - 1)  # use nc cells for mass integration
+            nc = int(10000 ** (1.0 / ndim) + 0.5)  # always use ~ 10000 mesh points
             dx = boxsize / nc
 
             #  integrate total mass in box
@@ -387,6 +423,9 @@ class ParticleGenerator(object):
             ip.boxsize_for_tree = self.boxsize_to_use[:ndim]
         else:
             ip.boxsize_for_tree = None
+
+        # set up stats
+        self.stats = IterStats(self.runparams.iter_max)
 
         # take note that we did this
         self._set_up = True
@@ -651,7 +690,7 @@ class ParticleGenerator(object):
 
         if dump_now or redistribute:
 
-            # first build new tree
+            # first build new tree, get smoothing lengths and densities
             h, rho = self.compute_h_and_rho()
 
             if dump_now:
@@ -669,13 +708,19 @@ class ParticleGenerator(object):
         # compute MODEL smoothing lengths
         oneoverrho = 1.0 / rho_model
         oneoverrhosum = np.sum(oneoverrho)
+        vol = 0.0
+        for d in range(ndim):
+            vol *= boxsize[d]
 
         if ndim == 1:
             hmodel = 0.5 * ipars.Nngb * oneoverrho / oneoverrhosum
+            #  hmodel = 0.5 * ipars.Nngb * oneoverrho / oneoverrhosum * vol
         elif ndim == 2:
             hmodel = np.sqrt(ipars.Nngb / np.pi * oneoverrho / oneoverrhosum)
+            #  hmodel = np.sqrt(ipars.Nngb / np.pi * oneoverrho / oneoverrhosum * vol)
         else:
             hmodel = np.cbrt(ipars.Nngb * 3 / 4 / np.pi * oneoverrho / oneoverrhosum)
+            #  hmodel = np.cbrt(ipars.Nngb * 3 / 4 / np.pi * oneoverrho / oneoverrhosum * vol)
 
         # init delta_r array
         delta_r = np.zeros(self.x.shape, dtype=np.float)
@@ -712,6 +757,7 @@ class ParticleGenerator(object):
                 delrsq += delta_r[:, d] ** 2
             delrsq = np.sqrt(delrsq)
[…]
