### 2020-07-29 f575a9e
renamed functions
