### 2020-07-29 4e16732
docstrings - read'em and weep!
