+++ swiftsimio/initial_conditions/generate_particles.py
@@ -1,4 +1,5 @@
-"""Generate SPH initial conditions for SPH simulations iteratively for a given density function following Arth et al.
+"""
+Generate SPH initial conditions for SPH simulations iteratively for a given density function following Arth et al.
 2019 (https://arxiv.org/abs/1907.11250).
 
 """
@@ -15,67 +16,84 @@ from swiftsimio import Writer
 
 class RunParams(object):
     r"""
-    TODO: dox
+    A class to store iteration runtime parameters. Before running the iteration,
+    these parameters are intended to be direcly accessed by the user.
 
 
     Attributes
     ---------------
-    iter_max: int, optional
-        max numbers of iterations for generating IC conditions
+    self.iter_max: int
+        maximal numbers of iterations to run for generating IC conditions
 
-    converge_thresh: float, optional
-        if enough particles are displaced by distance below threshold * mean
-        interparticle distance, stop iterating.
+    self.iter_min: int
+        minimal numbers of iterations to run for generating IC conditions
 
-    tolerance_part: float, optional
+    self.converge_thresh: float
+        upper threshold for what displacement is supposed to be considered as
+        converged. If enough particles are displaced by distance below
+        ``self.converge_thresh * mean interparticle distance``, stop iterating.
+        ``self.tolerance_part`` defines, what "enough particles" is.
+
+    self.tolerance_part: float
         tolerance for not converged particle fraction: this fraction of
-        particles can be displaced with distances > threshold
+        particles can be displaced with distances > self.converge_thresh
 
-    displacement_threshold: float, optional
-        iteration halt criterion: Don't stop until every particle is displaced
-        by distance < threshold * mean interparticle distance
+    self.displ_thresh: float
+        displacement threshold. It's an iteration continuation criterion: Don't
+        stop until every particle is displaced by a distance < `` self.displ_thresh
+        * mean interparticle distance``
 
-    delta_init: float or None, optional
+    self.delta_init: float or None
         initial normalization constant for particle motion in units of mean
-        interparticle distance. If ``None`` (default), delta_init will be set
+        interparticle distance. If ``None`` (default), ``self.delta_init`` will be set
         such that the maximal displacement found in the first iteration is
         normalized to 1 mean interparticle distance.
 
-    delta_reduction_factor: float, optional
-        multiply the normalization constant for particle motion by this factor
+    self.delta_reduct: float
+        normalization constant reduction factor. Intended to be > 0 and < 1.
+        Multiply the normalization constant for particle motion by this factor
         after every iteration. In certain difficult cases this might help the
         generation to converge if set to < 1.
 
-    delta_min: float, optional
+    self.delta_min: float
         minimal normalization constant for particle motion in units of mean
         interparticle distance.
 
-    redistribute_frequency: int, optional
-        redistribute a handful of particles every ``redistribute_frequency``
+    self.redist_freq: int
+        particle redistribution frequency.
+        Redistribute a handful of particles every ``self.redistribute_frequency``
         iteration. How many particles are redistributed is controlled with the
-        ``redistribute_fraction`` parameter.
+        ``self.redistribute_frac`` parameter.
 
-    redistribute_fraction: float, optional
+    self.redist_frac: float
         fraction of particles to be redistributed when doing so.
 
-    redistribute_fraction_reduction: float, optional
-        multiply the ``redistribute_fraction`` parameter by this factor every
+    self.redist_reduct: float
+        multiply the ``self.redistribute_fraction`` parameter by this factor every
[…]
