### 2020-07-29 b9ca5ff
forgot kernel data in __init__
