+++ swiftsimio/initial_conditions/__init__.py
@@ -3,3 +3,4 @@ Initial conditions generation.
 """
 
 from .generate_particles import ParticleGenerator
+from .IC_kernel import get_kernel_data
