### 2020-07-31 3bced0d
Address 2: used star_formation_rates in backend example which was empty
