### 2020-07-31 c929b13
Point 3: typo in rotation code.
