### 2020-07-31 21c841e
Address 4: require numba>=0.50.0 for install
