### 2020-07-31 e44bb40
Address point 5: typo in py-sphviewer docs
