+++ docs/source/visualisation/py-sphviewer.rst
@@ -15,12 +15,12 @@ You can get access to the objects through a sub-module as follows:
 
    resolution = 2048
 
-   gas = SPHViewer(data.gas, smooth_over="masses")
-   gas_temp = SPHViewer(
+   gas = SPHViewerWrapper(data.gas, smooth_over="masses")
+   gas_temp = SPHViewerWrapper(
        data.gas,
        smooth_over=data.gas.masses * data.gas.temperatures
    )
-   dark_matter = SPHViewer(data.dark_matter, smooth_over="masses")
+   dark_matter = SPHViewerWrapper(data.dark_matter, smooth_over="masses")
 
    gas.quick_view(xsize=resolution, ysize=resolution, r="infinity")
    gas_temp.quick_view(xsize=resolution, ysize=resolution, r="infinity")
