### 2020-07-31 0ac8053
Address 2: more info on halo catalogue integration.
