### 2020-07-31 SWIFTSIM/swiftsimio#67 @JBorrow
Hi, thanks for your kind words and detailed review! I'm especially glad that you found the API intuitive, as that was the major goal when it came to the design.

To address the points in your review:

1. I have added a whole section on this, which loads a halo from a catalogue and makes images of it. Hopefully the inline images give people a good idea of what's going on, too. The reason this wasn't done before is that I am not fully happy with the VR library API so it's likely to change in the near future, but you're right that this needed more documentation. That's available here: https://swiftsimio.readthedocs.io/en/latest/velociraptor/index.html.

2. Fixed, I've changed this to `entropies`.

3. Thanks for the catch, I have fixed that.

4. This is a tricky one; I have been trying for a long while to not require `numba` on install, but now I think we've crossed the threshold where it's used too much throughout the code that it really needs to be there. Alongside this, the `numba` API is not very stable, so we fix `swiftsimio` when they update things but then lose backwards compatibility. I have now required `numba>=0.50.0`.

5. Thanks, fixed.

6. Thanks, fixed.

Thanks again for taking the time to do this review! I will push out v4.0.3 with your comments addressed.
