+++ swiftsimio/initial_conditions/generate_particles.py
@@ -8,10 +8,11 @@ import unyt
 from math import erf
 from typing import Union
 import warnings
+import h5py
 
 from .IC_kernel import get_kernel_data
 from swiftsimio.optional_packages import KDTree, TREE_AVAILABLE
-from swiftsimio import Writer
+from swiftsimio import Writer, load
 
 
 class RunParams(object):
@@ -63,7 +64,8 @@ class RunParams(object):
         particle redistribution frequency.
         Redistribute a handful of particles every ``self.redistribute_frequency``
         iteration. How many particles are redistributed is controlled with the
-        ``self.redistribute_frac`` parameter.
+        ``self.redistribute_frac`` parameter. If = 0, no redistributions will be
+        performed.
 
     self.redist_frac: float
         fraction of particles to be redistributed when doing so.
@@ -78,7 +80,8 @@ class RunParams(object):
 
     self.dumpfreq: int
         frequency of dumps of the current state of the iteration. If set to zero,
-        no intermediate results will be stored.
+        no intermediate results will be stored. If > 0, it will also create a 
+        dump after the last iteration so a restart is possible.
 
     self.dump_basename: str
         Basename for intermediate dumps. The filename will be constructed as
@@ -370,6 +373,8 @@ class ParticleGenerator(object):
 
     # internal checks
     _set_up = False
+    _restart_finished = False
+    _restarting = False
 
     # unyt/result arrays
     coordinates: unyt.unyt_array
@@ -442,10 +447,10 @@ class ParticleGenerator(object):
         if not isinstance(unitsys, unyt.unit_systems.UnitSystem):
             raise TypeError("unitsys needs to be a unyt UnitSystem.")
 
-        if not isinstance(nx, int):
+        if not isinstance(nx, (int, np.integer)):
             raise TypeError("nx needs to be an integer")
 
-        if not isinstance(ndim, int):
+        if not isinstance(ndim, (int, np.integer)):
             raise TypeError("ndim needs to be an integer")
 
         self.rhofunc = rho
@@ -542,49 +547,54 @@ class ParticleGenerator(object):
         if not isinstance(res, np.ndarray):
             raise TypeError("rho(x, ndim) needs to return a numpy array as the result.")
 
-        # generate first positions if necessary
-        if x is None:
-            if method == "rejection":
-                self.coordinates = self.rejection_sample_coords()
-            elif method == "displaced":
-                self.coordinates = self.generate_displaced_uniform_coords(
-                    max_displ=max_displ
-                )
-            elif method == "uniform":
-                self.coordinates = self.generate_uniform_coords()
+        # if you are restarting, and did this already, don't do it again.
+        # this way, the user still can change parameters however they want
+        # and call self.initial_setup again.
+        if not self._restart_finished:
+
+            # generate first positions if necessary
+            if x is None:
+                if method == "rejection":
+                    self.coordinates = self.rejection_sample_coords()
+                elif method == "displaced":
+                    self.coordinates = self.generate_displaced_uniform_coords(
+                        max_displ=max_displ
+                    )
+                elif method == "uniform":
+                    self.coordinates = self.generate_uniform_coords()
+                else:
+                    raise ValueError("Unknown coordinate generation method:", method)
             else:
[…]
