+++ swiftsimio/initial_conditions/generate_particles.py
@@ -22,112 +22,114 @@ class RunParams(object):
 
 
     Attributes
-    ---------------
-    self.iter_max: int
+    ----------
+
+    max_iterations: int
         maximal numbers of iterations to run for generating IC conditions
 
-    self.iter_min: int
+    min_iterations: int
         minimal numbers of iterations to run for generating IC conditions
 
-    self.converge_thresh: float
+    convergence_threshold: float
         upper threshold for what displacement is supposed to be considered as
         converged. If enough particles are displaced by distance below
-        ``self.converge_thresh * mean interparticle distance``, stop iterating.
-        ``self.tolerance_part`` defines, what "enough particles" is.
+        ``self.convergence_threshold * mean interparticle distance``, stop 
+        iterating. ``self.unconverged_particle_number_tolerance`` defines, what
+        "enough particles" is.
 
-    self.tolerance_part: float
+    unconverged_particle_number_tolerance: float
         tolerance for not converged particle fraction: this fraction of
-        particles can be displaced with distances > self.converge_thresh
+        particles can be displaced with distances > self.convergence_threshold
 
-    self.displ_thresh: float
-        displacement threshold. It's an iteration continuation criterion: Don't
-        stop until every particle is displaced by a distance < `` self.displ_thresh
-        * mean interparticle distance``
+    displacement_threshold: float
+        It's an iteration continuation criterion: Don't
+        stop until every particle is displaced by a distance < 
+        ``self.displacement_threshold * mean interparticle distance``
 
-    self.delta_init: float or None
+    delta_init: float or None
         initial normalization constant for particle motion in units of mean
-        interparticle distance. If ``None`` (default), ``self.delta_init`` will be set
-        such that the maximal displacement found in the first iteration is
-        normalized to 1 mean interparticle distance.
+        interparticle distance. If ``None`` (default), ``self.delta_init`` will 
+        be set such that the maximal displacement found in the first iteration 
+        is normalized to 1 mean interparticle distance.
 
-    self.delta_reduct: float
+    delta_r_norm_reduction_factor: float
         normalization constant reduction factor. Intended to be > 0 and < 1.
         Multiply the normalization constant for particle motion by this factor
         after every iteration. In certain difficult cases this might help the
         generation to converge if set to < 1.
 
-    self.delta_min: float
+    min_delta_r_norm: float
         minimal normalization constant for particle motion in units of mean
         interparticle distance.
 
-    self.redist_freq: int
+    particle_redistribution_frequency: int
         particle redistribution frequency.
         Redistribute a handful of particles every ``self.redistribute_frequency``
         iteration. How many particles are redistributed is controlled with the
         ``self.redistribute_frac`` parameter. If = 0, no redistributions will be
         performed.
 
-    self.redist_frac: float
+    particle_redistribution_number_fraction: float
         fraction of particles to be redistributed when doing so.
 
-    self.redist_reduct: float
-        multiply the ``self.redistribute_fraction`` parameter by this factor every
-        time particles are being redistributed. In certain difficult cases this
-        might help the generation to converge if set to < 1.
+    particle_redistribution_number_reduction_factor: float
+        multiply the ``self.redistribute_fraction`` parameter by this factor 
+        every time particles are being redistributed. In certain difficult 
+        cases this might help the generation to converge if set to < 1.
 
[…]
