### 2020-07-31 SWIFTSIM/swiftsimio#67 @mtremmel
Thanks, this all looks good! This code is really neat, I was happy to have gotten a chance to try it out. I think you have addressed all of my concerns!
