+++ bibliography.bib
@@ -37,11 +37,19 @@
 }
 
 @misc{pynbody,
-  author = {{Pontzen}, A. and {Ro{\v s}kar}, R. and {Stinson}, G.~S. and {Woods},
-     R. and {Reed}, D.~M. and {Coles}, J. and {Quinn}, T.~R.},
-  title = "{pynbody: Astrophysics Simulation Analysis for Python}",
-  note = {Astrophysics Source Code Library, ascl:1305.002},
-  year = 2013
+       author = {{Pontzen}, Andrew and {Ro{\v{s}}kar}, Rok and {Stinson}, Greg and
+         {Woods}, Rory},
+        title = "{pynbody: N-Body/SPH analysis for python}",
+     keywords = {Software},
+         year = 2013,
+        month = may,
+          eid = {ascl:1305.002},
+        pages = {ascl:1305.002},
+archivePrefix = {ascl},
+       eprint = {1305.002},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013ascl.soft05002P},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+ howpublished = {\url{http://ascl.net/1305.002}}
 }
 
 @MISC{pnbody,
@@ -55,7 +63,8 @@
 archivePrefix = {ascl},
        eprint = {1302.004},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2013ascl.soft02004R},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System},
+ howpublished = {\url{http://ascl.net/1302.004}}
 }
 
 @book{h5py,
