+++ bibliography.bib
@@ -1,21 +1,23 @@
 
-@article{yt,
-	title = {yt: {A} {MULTI}-{CODE} {ANALYSIS} {TOOLKIT} {FOR} {ASTROPHYSICAL} {SIMULATION} {DATA}},
-	volume = {192},
-	issn = {0067-0049, 1538-4365},
-	shorttitle = {yt},
-	url = {https://iopscience.iop.org/article/10.1088/0067-0049/192/1/9},
-	doi = {10.1088/0067-0049/192/1/9},
-	abstract = {The analysis of complex multiphysics astrophysical simulations presents a unique and rapidly growing set of challenges: reproducibility, parallelization, and vast increases in data size and complexity chief among them. In order to meet these challenges, and in order to open up new avenues for collaboration between users of multiple simulation platforms, we present yt (available at http://yt.enzotools.org/) an open source, communitydeveloped astrophysical analysis and visualization toolkit. Analysis and visualization with yt are oriented around physically relevant quantities rather than quantities native to astrophysical simulation codes. While originally designed for handling Enzo’s structure adaptive mesh reﬁnement data, yt has been extended to work with several different simulation methods and simulation codes including Orion, RAMSES, and FLASH. We report on its methods for reading, handling, and visualizing data, including projections, multivariate volume rendering, multidimensional histograms, halo ﬁnding, light cone generation, and topologically connected isocontour identiﬁcation. Furthermore, we discuss the underlying algorithms yt uses for processing and visualizing data, and its mechanisms for parallelization of analysis tasks.},
-	language = {en},
-	number = {1},
-	urldate = {2020-04-17},
-	journal = {The Astrophysical Journal Supplement Series},
-	author = {Turk, Matthew J. and Smith, Britton D. and Oishi, Jeffrey S. and Skory, Stephen and Skillman, Samuel W. and Abel, Tom and Norman, Michael L.},
-	month = jan,
-	year = {2011},
-	pages = {9},
-	file = {Turk et al. - 2011 - yt A MULTI-CODE ANALYSIS TOOLKIT FOR ASTROPHYSICA.pdf:/Users/mphf18/Zotero/storage/NRC9WYWY/Turk et al. - 2011 - yt A MULTI-CODE ANALYSIS TOOLKIT FOR ASTROPHYSICA.pdf:application/pdf}
+@ARTICLE{yt,
+       author = {{Turk}, Matthew J. and {Smith}, Britton D. and {Oishi}, Jeffrey S. and
+         {Skory}, Stephen and {Skillman}, Samuel W. and {Abel}, Tom and
+         {Norman}, Michael L.},
+        title = "{yt: A Multi-code Analysis Toolkit for Astrophysical Simulation Data}",
+      journal = {\apjs},
+     keywords = {cosmology: theory, methods: data analysis, methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2011,
+        month = jan,
+       volume = {192},
+       number = {1},
+          eid = {9},
+        pages = {9},
+          doi = {10.1088/0067-0049/192/1/9},
+archivePrefix = {arXiv},
+       eprint = {1011.3514},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2011ApJS..192....9T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
 @article{Gadget2,
@@ -83,7 +85,7 @@ archivePrefix = {ascl},
   number = {28},
   pages = {809},
[…]
