+++ bibliography.bib
@@ -4,7 +4,7 @@
          {Skory}, Stephen and {Skillman}, Samuel W. and {Abel}, Tom and
          {Norman}, Michael L.},
         title = "{yt: A Multi-code Analysis Toolkit for Astrophysical Simulation Data}",
-      journal = {\apjs},
+      journal = {The Astrophysical Journal Supplement Series},
      keywords = {cosmology: theory, methods: data analysis, methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2011,
         month = jan,
