### 2020-07-31 #2430 @dfm
@JBorrow: I opened a pull request with some edits for the `.bib` file - the manuscript otherwise looks great to me. After you merge that, can you build the paper here to make sure you're happy with the proofs (it will be harder to make too many changes later!)

Then, once you're happy with the proofs, can you bump the version number of the code and archive it (using Zenodo or similar). The archive should have the new version number and it should also have *exactly* the same title and author list as your paper (with Zenodo you can edit the metadata after the DOI is minted).

Finally, please report the new version number and the DOI for the archive in this thread.

Thanks!
