### 2020-08-01 1ea18b3
Revved to v4.0.4 for JOSS
