### 2020-08-01 #2430 @dfm
@JBorrow: It looks like all we got in that archive is the wheel of the Python package. It should be the full source of the repo instead!
